% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{actionability_doc_note}
\alias{actionability_doc_note}
\title{Get documentation string for AMP/ASCO/CAP clinical actionability tiers}
\usage{
actionability_doc_note()
}
\value{
A documentation string
}
\description{
Get documentation string for AMP/ASCO/CAP clinical actionability tiers
}
