% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_gwas_citation_phenotype}
\alias{add_gwas_citation_phenotype}
\title{Function that adds GWAS citation/phenotype to GWAS hit found through PCGR annotation}
\usage{
add_gwas_citation_phenotype(vcf_data_df, pcgr_data,
  p_value_threshold = 5e-08)
}
\arguments{
\item{vcf_data_df}{Data frame of sample variants from VCF}

\item{pcgr_data}{PCGR data structure}
}
\value{
vcf_data_df
}
\description{
Function that adds GWAS citation/phenotype to GWAS hit found through PCGR annotation
}
