% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_pfam_domain_links}
\alias{add_pfam_domain_links}
\title{Function that adds PFAM name descriptions to PFAM identifiers}
\usage{
add_pfam_domain_links(vcf_data_df, pcgr_data)
}
\arguments{
\item{vcf_data_df}{}
}
\value{
vcf_data_df_pfam
}
\description{
Function that adds PFAM name descriptions to PFAM identifiers
}
