% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_read_support}
\alias{add_read_support}
\title{Function that adds read support (depth, allelic fraction) for tumor and normal}
\usage{
add_read_support(vcf_data_df, pcgr_config)
}
\arguments{
\item{vcf_data_df}{data frame with variants}

\item{pcgr_config}{list with configuration values}
}
\value{
vcf_data_df
}
\description{
Function that adds read support (depth, allelic fraction) for tumor and normal
}
