% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_swissprot_feature_descriptions}
\alias{add_swissprot_feature_descriptions}
\title{Function that adds SwissProt feature descriptions based on keys coming from pcgr pipeline}
\usage{
add_swissprot_feature_descriptions(vcf_data_df, pcgr_data)
}
\arguments{
\item{vcf_data_df}{}
}
\value{
vcf_data_df
}
\description{
Function that adds SwissProt feature descriptions based on keys coming from pcgr pipeline
}
