% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{append_alteration_name}
\alias{append_alteration_name}
\title{Function that appends informative alteration names to
sample variants based on gene symbol, consequence and
HGVS annotations}
\usage{
append_alteration_name(var_data_df = NULL)
}
\arguments{
\item{var_data_df}{Data frame of sample variants from VCF}
}
\value{
var_data_df
}
\description{
Function that appends informative alteration names to
sample variants based on gene symbol, consequence and
HGVS annotations
}
