% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_annotation_links}
\alias{append_annotation_links}
\title{Function that appends multiple HTML annotation links to variant identifiers
e.g. COSMIC, CLINVAR, REFSEQ etc}
\usage{
append_annotation_links(vcf_data_df, skip = NULL)
}
\arguments{
\item{vcf_data_df}{data frame with variant entries}

\item{skip}{elements to be ignored during annotation}
}
\description{
Function that appends multiple HTML annotation links to variant identifiers
e.g. COSMIC, CLINVAR, REFSEQ etc
}
