% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{append_cancer_association_ranks}
\alias{append_cancer_association_ranks}
\title{Function that appends cancer gene evidence links}
\usage{
append_cancer_association_ranks(
  var_data_df = NULL,
  ref_data = NULL,
  primary_site = "Any",
  pos_var = "POS"
)
}
\arguments{
\item{var_data_df}{Data frame of sample variants from VCF}

\item{ref_data}{PCGR reference data bundle object}

\item{primary_site}{Primary tumor site}

\item{pos_var}{variable reflecting chromosome order (POS/SEGMENT_START)}
}
\value{
vcf_data_df
}
\description{
Function that appends cancer gene evidence links
}
