% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_dbmts_var_link}
\alias{append_dbmts_var_link}
\title{Function that adds HTML links to different genetic variant identifiers}
\usage{
append_dbmts_var_link(var_df)
}
\arguments{
\item{var_df}{data frame with variants}

\item{linktype}{type of link}

\item{pcgr_data}{PCGR data structure}
}
\description{
Function that adds HTML links to different genetic variant identifiers
}
