% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{append_dbnsfp_var_link}
\alias{append_dbnsfp_var_link}
\title{Function that assigns HTML links to dbNSFP prediction entries}
\usage{
append_dbnsfp_var_link(var_df)
}
\arguments{
\item{var_df}{data frame with variant entries}
}
\value{
var_df
}
\description{
Function that assigns HTML links to dbNSFP prediction entries
}
