% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_drug_var_link}
\alias{append_drug_var_link}
\title{Function that adds HTML links to different genetic variant identifiers}
\usage{
append_drug_var_link(vcf_data_df, ref_data = NULL)
}
\arguments{
\item{vcf_data_df}{data frame with variants}

\item{ref_data}{PCGR/CPSR reference data object}
}
\description{
Function that adds HTML links to different genetic variant identifiers
}
