% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_otargets_pheno_link}
\alias{append_otargets_pheno_link}
\title{Function that adds HTML links to different genetic variant identifiers}
\usage{
append_otargets_pheno_link(
  var_df,
  pcgr_data = NULL,
  oncotree = NULL,
  linktype = "dbsource"
)
}
\arguments{
\item{var_df}{data frame with variants}

\item{pcgr_data}{PCGR data structure}

\item{oncotree}{Oncotree data frame}

\item{linktype}{type of link}
}
\description{
Function that adds HTML links to different genetic variant identifiers
}
