% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_pfeature_descriptions}
\alias{append_pfeature_descriptions}
\title{Function that adds SwissProt feature descriptions based on
key identifiers coming from PCGR pipeline}
\usage{
append_pfeature_descriptions(vcf_data_df, feature_descriptions)
}
\arguments{
\item{vcf_data_df}{Data frame of sample variants from VCF}

\item{feature_descriptions}{Data frame with SwissProt feature descriptions}
}
\value{
vcf_data_df
}
\description{
Function that adds SwissProt feature descriptions based on
key identifiers coming from PCGR pipeline
}
