% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{append_protein_domains}
\alias{append_protein_domains}
\title{Function that adds protein domain annotations (PFAM)
for a list of variants and associated targeted}
\usage{
append_protein_domains(var_df, ref_data = NULL)
}
\arguments{
\item{var_df}{data frame with variants}

\item{ref_data}{PCGR/CPSR reference data object}
}
\value{
var_df
}
\description{
Function that adds protein domain annotations (PFAM)
for a list of variants and associated targeted
}
