% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{append_styled_cna_vclass}
\alias{append_styled_cna_vclass}
\title{Function that styles CNA variant classes (amplification,
gain, hetdel, homdel/hemdel) with more informative names
for display in PCGR reports and other outputs}
\usage{
append_styled_cna_vclass(var_data_df = NULL, colname = "VARIANT_CLASS_STYLED")
}
\arguments{
\item{var_data_df}{Data frame of CNA calls}

\item{colname}{Name of column to be styled and appended to var_data_df}
}
\value{
var_data_df
}
\description{
Function that styles CNA variant classes (amplification,
gain, hetdel, homdel/hemdel) with more informative names
for display in PCGR reports and other outputs
}
