% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{append_tfbs_annotation}
\alias{append_tfbs_annotation}
\title{Function that adds TFBS annotations (dbMTS) to genetic variant identifiers}
\usage{
append_tfbs_annotation(var_df)
}
\arguments{
\item{var_df}{data frame with variants}
}
\description{
Function that adds TFBS annotations (dbMTS) to genetic variant identifiers
}
