% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_ucsc_segment_link}
\alias{append_ucsc_segment_link}
\title{Function that appends a link to UCSC for a genomic segment}
\usage{
append_ucsc_segment_link(
  var_df,
  hgname = "hg38",
  chrom = NULL,
  start = NULL,
  end = NULL
)
}
\arguments{
\item{var_df}{data frame with genomic variants}

\item{hgname}{name of genoome assembly ('hg38','hg19')}

\item{chrom}{chromosome name}

\item{start}{chromosome start coordinate}

\item{end}{chromosome end coordinate}
}
\value{
var_df
}
\description{
Function that appends a link to UCSC for a genomic segment
}
