% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_classification.R
\name{assign_bm_tier_support_ttspecific}
\alias{assign_bm_tier_support_ttspecific}
\title{For all biomarker evidence items with specified confidence level,
assign whether these are tier-defining or providing additional
support - tumor-type specific query}
\usage{
assign_bm_tier_support_ttspecific(
  variants_tier_classified = NULL,
  vartype = NULL,
  primary_site = "Lung",
  etype_for_tiering = c("predictive"),
  biomarker_items = NULL
)
}
\arguments{
\item{variants_tier_classified}{data frame with variant tier classifications
based on support from biomarker evidence item}

\item{vartype}{variant type (e.g. 'snv_indel', 'cna', 'fusion')}

\item{primary_site}{primary tumor site of query tumor (e.g. 'Lung',
'Breast', 'Any')}

\item{etype_for_tiering}{evidence type(s) to consider for tier classification}

\item{biomarker_items}{data frame with biomarker evidence items}
}
\value{
data frame with biomarker evidence items classified as tier-defining
or providing additional support
}
\description{
The function considers the strength of evidence (evidence levels) and
the match between biomarker site and primary site of query tumor.
Mark evidence considered tier-defining, while other biomarker evidence items
as additional support
}
