% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsr.R
\name{assign_cpsr_tier}
\alias{assign_cpsr_tier}
\title{Function that assign variants to different tiers for
prioritization of germline variants}
\usage{
assign_cpsr_tier(cpg_calls, cpsr_config, display_tags)
}
\arguments{
\item{cpg_calls}{data frame with variants in predisposition_genes}
}
\description{
Function that assign variants to different tiers for
prioritization of germline variants
}
