% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{assign_mutation_type}
\alias{assign_mutation_type}
\title{Function that assigns one of six mutation types to a list of mutations}
\usage{
assign_mutation_type(var_df)
}
\arguments{
\item{var_df}{data frame with variants}
}
\value{
var_df
}
\description{
Function that assigns one of six mutation types to a list of mutations
}
