% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsr.R
\name{assign_pathogenicity_evidence}
\alias{assign_pathogenicity_evidence}
\title{Function that assigns variant pathogenicity evidence based on ACMG guidelines}
\usage{
assign_pathogenicity_evidence(cpg_calls, cpsr_config, pcgr_data)
}
\arguments{
\item{cpg_calls}{sample calls with dbnsfp annotations}

\item{cpsr_config}{cpsr configuration object}

\item{pcgr_data}{pcgr data object}
}
\value{
cpg_calls
}
\description{
Function that assigns variant pathogenicity evidence based on ACMG guidelines
}
