% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{assign_poplevel_frequency_class}
\alias{assign_poplevel_frequency_class}
\title{Function that assigns a category ('Rare','Common' etc) to population-specific germline frequencies}
\usage{
assign_poplevel_frequency_class(var_df, pop_af_column)
}
\arguments{
\item{var_df}{data frame with variants}

\item{pop_af_column}{population_column}
}
\value{
var_df
}
\description{
Function that assigns a category ('Rare','Common' etc) to population-specific germline frequencies
}
