% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acmg.R
\name{assign_tier1_tier2_acmg}
\alias{assign_tier1_tier2_acmg}
\title{Function that assigns evidence items for SNVs/InDels to ACMG tiers 1 and 2}
\usage{
assign_tier1_tier2_acmg(pcg_report_snv_indel)
}
\arguments{
\item{pcg_report_snv_indel}{report object for snv/indels}
}
\value{
pcg_report_snv_indel data frame with all report elements
}
\description{
Function that assigns evidence items for SNVs/InDels to ACMG tiers 1 and 2
}
