% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsr.R
\name{assign_variant_tiers}
\alias{assign_variant_tiers}
\title{Function that assign variants to different tiers for
prioritization of germline variants}
\usage{
assign_variant_tiers(
  cpg_calls,
  config = NULL,
  cpsr_display_cols = NULL,
  cpsr_tsv_cols = NULL
)
}
\arguments{
\item{cpg_calls}{data frame with variants in predisposition_genes}

\item{config}{CPSR configuration object with run settings}

\item{cpsr_display_cols}{list of variables for display in report}

\item{cpsr_tsv_cols}{list of variables for display in report}
}
\description{
Function that assign variants to different tiers for
prioritization of germline variants
}
