% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_classification.R
\name{assign_variant_tiers_fusion}
\alias{assign_variant_tiers_fusion}
\title{Assign tiers of clinical significance (AMP/ASCO/CAP framework) to
RNA fusions}
\usage{
assign_variant_tiers_fusion(
  primary_site = "Any",
  biomarker_mapping_confidence = "medium",
  var_df = NULL,
  etype_for_tiering = c("predictive"),
  biomarker_items = NULL
)
}
\arguments{
\item{primary_site}{primary tumor site of query tumor
(e.g. 'Lung', 'Breast', 'Any')}

\item{biomarker_mapping_confidence}{confidence level of variant-biomarker mapping
resolution  (e.g. 'high' or 'medium')}

\item{var_df}{data frame with somatic RNA fusions}

\item{etype_for_tiering}{evidence type(s) to consider for tier classification
(e.g. 'predictive', 'prognostic', 'diagnostic')}

\item{biomarker_items}{data frame with biomarker evidence items}
}
\value{
data frame with tier classifications for somatic RNA fusions
based on biomarker evidence items and variant properties
}
\description{
Function that assigns tiers of clinical significance (AMP/ASCO/CAP
framework) to RNA fusions based on biomarker evidence items.
The function considers the strength of evidence (evidence levels)
and the match between biomarker site and primary site of query tumor.
The function also considers oncogene properties of fusion partners to
assign tier 3 status to variants with uncertain clinical significance.
}
