% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_classification.R
\name{assign_variant_top_tiers_ttagnostic}
\alias{assign_variant_top_tiers_ttagnostic}
\title{Assign top tiers of clinical significance (AMP/ASCO/CAP framework) to
variants based on data from biomarker evidence items (tumor-type agnostic query)}
\usage{
assign_variant_top_tiers_ttagnostic(
  biomarker_items = NULL,
  biomarker_mapping_confidence = "high",
  var_df = NULL,
  etype_for_tiering = c("predictive")
)
}
\arguments{
\item{biomarker_items}{data frame with biomarker evidence items}

\item{biomarker_mapping_confidence}{confidence level of variant-biomarker mapping
resolution  (e.g. 'high' or 'medium')}

\item{var_df}{data frame with variants (SNVs/InDels, CNAs, fusions)}

\item{etype_for_tiering}{evidence type(s) to consider for tier classification
(e.g. 'predictive', 'prognostic', 'diagnostic')}
}
\value{
data frame with top tier classifications for variants based on biomarker evidence items
}
\description{
The function considers the strength of evidence (evidence levels) and
the match between biomarker site and primary site of query tumor.
}
