% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\name{bp_junction_transcript_overlap}
\alias{bp_junction_transcript_overlap}
\title{Find transcripts covering given splice junction breakpoints}
\usage{
bp_junction_transcript_overlap(bp_junctions = NULL, ref_data = NULL)
}
\arguments{
\item{bp_junctions}{A data frame with columns:
\itemize{
\item BP_CHROM: Chromosome of the breakpoint junction
\item BP_POSITION: Chromosome position of the breakpoint junction
}}

\item{ref_data}{PCGR reference data bundle (list)}
}
\value{
A data frame with columns:
\itemize{
\item CHROM: Chromosome of the breakpoint junction
\item BP_POSITION: Position of the breakpoint junction
\item ENSEMBL_TRANSCRIPT_ID: Ensembl transcript ID covering the splice junction
\item ENSEMBL_GENE_ID: Ensembl gene ID
\item GENE_BIOTYPE: Biotype of the transcript
\item TRANSCRIPT_START: Start position of the transcript
\item TRANSCRIPT_END: End position of the transcript
}
}
\description{
This function identifies transcripts that cover specified
junction breakpoints.
}
