% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer_phenotypes_regex}
\alias{cancer_phenotypes_regex}
\title{Regular expression of terms indicative of cancer-related phenotypes and syndromes}
\format{
A long regular expression of cancer-related phenotype terms
}
\usage{
cancer_phenotypes_regex
}
\description{
Regular expression of terms indicative of cancer-related phenotypes and syndromes
}
\keyword{datasets}
