% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Function that checks the existence of a file}
\usage{
check_file_exists(fname)
}
\arguments{
\item{fname}{Name of file to check}
}
\description{
Function that checks the existence of a file
}
