% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{clean_gnomad_annotations}
\alias{clean_gnomad_annotations}
\title{Clean gnomAD VCF annotations}
\usage{
clean_gnomad_annotations(var_df = NULL)
}
\arguments{
\item{var_df}{data frame with variants}
}
\value{
var_df with cleaned gnomAD annotations
}
\description{
Function that cleans and processes gnomAD annotations in the
raw VCF-annotated callset for SNVs/InDels
}
