% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{clinvar_germline_status}
\alias{clinvar_germline_status}
\title{Function that assigns a logical to STATUS_CLINVAR_GERMLINE based on
whether a ClinVar entry of germline origin is found for a given variant
(for entries in a data frame)}
\usage{
clinvar_germline_status(sample_calls)
}
\arguments{
\item{sample_calls}{data frame with sample calls}
}
\description{
Function that assigns a logical to STATUS_CLINVAR_GERMLINE based on
whether a ClinVar entry of germline origin is found for a given variant
(for entries in a data frame)
}
