% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna.R
\name{cna_segment_annotation}
\alias{cna_segment_annotation}
\title{Function that annotates CNV segment files}
\usage{
cna_segment_annotation(cna_file, logR_gain, logR_homdel, pcgr_data,
  transcript_overlap_frac = 100)
}
\arguments{
\item{cna_file}{CNV file name}

\item{logR_gain}{logR treshold for annotating copy number amplifications}

\item{logR_homdel}{logR threshold for annotating homozygous deletions}

\item{pcgr_data}{List of data frames with PCGR data annotations}
}
\value{
cna_data
}
\description{
Function that annotates CNV segment files
}
