% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{color_palette}
\alias{color_palette}
\title{Color encodings for report elements of PCGR/CPSR}
\format{
A list object with different report elements that are color-coded in
PCGR/CPSR reports. Each list element have two vectors: 'levels' and 'values'.
Currently, the following list elements are included:
\itemize{
\item \emph{pathogenicity} - Colors for five-level pathogenicity levels (CPSR)
\item \emph{clinical_evidence} - Colors for strength of evidence of cancer-variant associations (A-E)
\item \emph{tier} - Colors for tier levels for variant prioritization (PCGR)
\item \emph{report_color} - Colors for PCGR assay mode (tumor-control vs. tumor-only)
\item \emph{warning} - Color for warning (low confidence in PCGR analysis output)
\item \emph{success} - Color for success (no evident uncertainty in PCGR analysis output)
}
}
\usage{
color_palette
}
\description{
Color encodings for report elements of PCGR/CPSR
}
\keyword{datasets}
