% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cosmic_sbs_signatures}
\alias{cosmic_sbs_signatures}
\title{List of COSMIC reference mutational signatures (SBS, v3.4)}
\format{
A list with two matrix objects ('all' and 'no_artefacts').
One matrix contains the COSMIC reference mutational signatures without signature
artefacts ('no_artefacts', number of columns = 68), while the other contains
all signatures, including artefacts ('all', number of columns = 86). Each
matrix has 96 rows, one for each of the 96 possible trinucleotide contexts.
}
\usage{
cosmic_sbs_signatures
}
\description{
List of COSMIC reference mutational signatures (SBS, v3.4)
}
\keyword{datasets}
