% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{cosmic_somatic_status}
\alias{cosmic_somatic_status}
\title{Function that assigns a logical (STATUS_COSMIC) reflecting whether
a variant co-incides with an entry in COSMIC (germline)}
\usage{
cosmic_somatic_status(sample_calls)
}
\arguments{
\item{sample_calls}{data frame with sample variant calls}
}
\description{
Function that assigns a logical (STATUS_COSMIC) reflecting whether
a variant co-incides with an entry in COSMIC (germline)
}
