% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsr.R
\name{custom_bed_genes}
\alias{custom_bed_genes}
\title{Function that reports protein-coding geneset that overlaps BED file}
\usage{
custom_bed_genes(bed_file, pcgr_data)
}
\arguments{
\item{bed_file}{BED file name with selected transcripts from panel 0}

\item{pcgr_data}{object with PCGR annotation data}
}
\description{
Function that reports protein-coding geneset that overlaps BED file
}
