% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_coltype_defs}
\alias{data_coltype_defs}
\title{List of coltype definitions for input files to pcgrr (e.g. VCF-converted TSV, CNA TVS etc.)}
\format{
An object of class \code{list} of length 3.
}
\usage{
data_coltype_defs
}
\description{
List of coltype definitions for input files to pcgrr (e.g. VCF-converted TSV, CNA TVS etc.)
}
\keyword{datasets}
