% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_integrity_check}
\alias{data_integrity_check}
\title{Function that performs a sanity check of query VCF for pcgrr package,
ensuring that all required tags are present in annotated VCF}
\usage{
data_integrity_check(vcf_df, pcgr_data, config = NULL, workflow = "pcgr")
}
\arguments{
\item{vcf_df}{data frame with variants from VCF}

\item{pcgr_data}{PCGR data object}

\item{config}{PCGR/CPSR config object}

\item{workflow}{type of workflow (pcgr/cpsr)}
}
\description{
Function that performs a sanity check of query VCF for pcgrr package,
ensuring that all required tags are present in annotated VCF
}
