% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsr.R
\name{detect_cancer_traits_clinvar}
\alias{detect_cancer_traits_clinvar}
\title{Function that retrieves variants in cancer predisposition genes linked
to cancer-related conditions according to ClinVar}
\usage{
detect_cancer_traits_clinvar(cpg_calls, oncotree, umls_map)
}
\arguments{
\item{oncotree}{data frame with hereditary cancer phenotypes from OncoTree}

\item{umls_map}{}

\item{calls}{data frame with variants in predisposition_genes}
}
\description{
Function that retrieves variants in cancer predisposition genes linked
to cancer-related conditions according to ClinVar
}
