% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_vcf_sample_name}
\alias{detect_vcf_sample_name}
\title{A function that detects whether the sample name in
variant data frame is unique (as present in column name VCF_SAMPLE_ID), throws an error if
multiple sample names are present for the CPSR workflow}
\usage{
detect_vcf_sample_name(df, sample_name = NULL, cpsr = FALSE)
}
\arguments{
\item{df}{VCF data frame}

\item{sample_name}{name of sample identifier}

\item{cpsr}{logical indicating CPSR workflow}
}
\value{
df Vranges object
}
\description{
A function that detects whether the sample name in
variant data frame is unique (as present in column name VCF_SAMPLE_ID), throws an error if
multiple sample names are present for the CPSR workflow
}
