% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsr.R
\name{determine_pathogenicity_classification}
\alias{determine_pathogenicity_classification}
\title{Function that assigns final pathogenicity classification (B, LB, VUS, P, LP)
based on accumulated scores from different ACMG criteria and pre-defined
cutoffs (calibrated against ClinVar)}
\usage{
determine_pathogenicity_classification(cpg_calls)
}
\arguments{
\item{cpg_calls}{data frame with variant calls in predisposition genes}
}
\value{
cpg_calls data frame with pathogenicity classification appended
}
\description{
Function that assigns final pathogenicity classification (B, LB, VUS, P, LP)
based on accumulated scores from different ACMG criteria and pre-defined
cutoffs (calibrated against ClinVar)
}
