% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_string_replace}
\alias{df_string_replace}
\title{Function that performs stringr::str_replace on strings of multiple
string columns of a dataframe}
\usage{
df_string_replace(df, strings, pattern, replacement, replace_all = F)
}
\arguments{
\item{df}{data frame}

\item{strings}{name of columns for which string replace is to be performed}

\item{pattern}{pattern to replace}

\item{replacement}{string to replace}

\item{replace_all}{logical - replace all occurrences}
}
\value{
df
}
\description{
Function that performs stringr::str_replace on strings of multiple
string columns of a dataframe
}
