% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{effect_prediction_algos}
\alias{effect_prediction_algos}
\title{List of URLs for a range of variant effect prediction algorithms}
\format{
A data.frame with 21 rows and 3 columns that indicates URL's for
variant effect prediction algorithms
\itemize{
\item \emph{algorithm} - Name encoding for effect prediction algorithm
\item \emph{url} - URL
\item \emph{display_name} - Display name for use in reporting
}
}
\usage{
effect_prediction_algos
}
\description{
List of URLs for a range of variant effect prediction algorithms
}
\keyword{datasets}
