% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exonic_filter_levels}
\alias{exonic_filter_levels}
\title{Exonic filter levels}
\format{
An object of class \code{character} of length 2.
}
\usage{
exonic_filter_levels
}
\description{
Exonic filter levels
}
\keyword{datasets}
