% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_quarto_evars}
\alias{export_quarto_evars}
\title{Export Quarto Environment Variables}
\usage{
export_quarto_evars(x)
}
\arguments{
\item{x}{Path to conda/envs/pcgrr/etc/conda/activate.d/quarto.sh}
}
\description{
Export quarto environment variables
required when using conda.
}
