% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{expression_doc_note}
\alias{expression_doc_note}
\title{Get documentation string for RNA expression analysis (outlier, similarity)}
\usage{
expression_doc_note()
}
\value{
A documentation string
}
\description{
Get documentation string for RNA expression analysis (outlier, similarity)
}
