% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncokb.R
\name{fetch_oncokb_cna_annotation}
\alias{fetch_oncokb_cna_annotation}
\title{Fetch OncoKB annotation for copy number alteration}
\usage{
fetch_oncokb_cna_annotation(
  hugo_symbol = NULL,
  cna_type = NULL,
  oncotree_code = NULL,
  oncokb_token = NULL
)
}
\arguments{
\item{hugo_symbol}{Gene symbol}

\item{cna_type}{Type of CNA: "Amplification" or "Deletion"}

\item{oncotree_code}{OncoTree code/name (e.g., "SKIN", "BREAST")}

\item{oncokb_token}{OncoKB API token}
}
\value{
List containing the complete JSON response from OncoKB API
}
\description{
Fetch OncoKB annotation for copy number alteration
}
