% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_db_germline_variants}
\alias{filter_db_germline_variants}
\title{Function that filters a data frame with variants according to population-specific germline frequencies}
\usage{
filter_db_germline_variants(var_df, pop = "EUR", dbquery = "1KG",
  max_tolerated_af = 0.01)
}
\arguments{
\item{var_df}{data frame with variants}

\item{pop}{population ('european' or 'global')}

\item{dbquery}{'1KG' or 'gnomAD'}

\item{min_af}{minimum allele frequency required for variant to be filtered}
}
\value{
var_df
}
\description{
Function that filters a data frame with variants according to population-specific germline frequencies
}
