% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{filter_eitems_by_site}
\alias{filter_eitems_by_site}
\title{Function that filters clinical evidence items by tumor type/primary site}
\usage{
filter_eitems_by_site(eitems = NULL, ontology = NULL, primary_site = "")
}
\arguments{
\item{eitems}{data frame with clinical evidence items}

\item{ontology}{phenotype ontology data frame}

\item{primary_site}{primary tumor site}
}
\description{
Function that filters clinical evidence items by tumor type/primary site
}
