% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{filter_maf_file}
\alias{filter_maf_file}
\title{Function that takes a MAF file generated with vcf2maf
and filters out variants that are presumably germline (tumor-only run)}
\usage{
filter_maf_file(callset, settings)
}
\arguments{
\item{callset}{Callset with pre-processed somatic SNV/InDel variants}

\item{settings}{PCGR run/configuration settings}
}
\description{
Function that takes a MAF file generated with vcf2maf
and filters out variants that are presumably germline (tumor-only run)
}
