% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{filter_no_control_calls}
\alias{filter_no_control_calls}
\title{Function that excludes germline variants from an annotated callset with mix of somatic and germline variants}
\usage{
filter_no_control_calls(sample_calls_df, sample_name = NULL,
  pcgr_config = NULL)
}
\arguments{
\item{sample_calls_df}{data frame with mix of germline + somatic mutations (run using tumor only)}

\item{pcgr_config}{list with PCGR configuration parameters}
}
\value{
sample_calls filtered call set
}
\description{
Function that excludes germline variants from an annotated callset with mix of somatic and germline variants
}
