% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_read_support}
\alias{filter_read_support}
\title{Function that filters variant set on (depth, allelic fraction)
for tumor and normal and filters according to settings}
\usage{
filter_read_support(vcf_df, config = NULL, precision = 3)
}
\arguments{
\item{vcf_df}{data frame with variants}

\item{config}{list with workflow configuration values}

\item{precision}{number of significant digits for allelic fraction estimation}
}
\value{
vcf_df
}
\description{
Function that filters variant set on (depth, allelic fraction)
for tumor and normal and filters according to settings
}
