% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{fusion_doc_note}
\alias{fusion_doc_note}
\title{Get documentation string for RNA fusion analysis}
\usage{
fusion_doc_note()
}
\value{
A documentation string
}
\description{
Get documentation string for RNA fusion analysis
}
