% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_annotation_link}
\alias{generate_annotation_link}
\title{A function that generates a HTML link for selected identifiers (DBSNP, COSMIC, CLINVAR, ENTREZ)}
\usage{
generate_annotation_link(df, vardb = "DBSNP", group_by_var = "VAR_ID",
  url_prefix = "http://www.ncbi.nlm.nih.gov/projects/SNP/snp_ref.cgi?rs=",
  link_key_var = "DBSNPRSID", link_display_var = "DBSNPRSID")
}
\arguments{
\item{df}{data frame}

\item{vardb}{type of database}

\item{group_by_var}{variable used for grouping (VAR_ID)}

\item{url_prefix}{url prefix for link generation}

\item{link_key_var}{variable used in url for linking}

\item{link_display_var}{variable used in url for display}
}
\value{
df_annotation_links
}
\description{
A function that generates a HTML link for selected identifiers (DBSNP, COSMIC, CLINVAR, ENTREZ)
}
