% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_biomarker_tsv}
\alias{generate_biomarker_tsv}
\title{Function that generates a data frame with basic biomarker annotations from tier1 variants}
\usage{
generate_biomarker_tsv(clinical_evidence_items, sample_name = "test",
  tier = "tier 1")
}
\arguments{
\item{tier1_variants}{df with tier 1 variants}

\item{sample_id}{Sample identifier}
}
\value{
tsv_variants data frame with all tier 1 biomarkers for tab-separated output
}
\description{
Function that generates a data frame with basic biomarker annotations from tier1 variants
}
