% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsr.R
\name{generate_cpsr_report}
\alias{generate_cpsr_report}
\title{Function that generates predisposition_report - CPSR}
\usage{
generate_cpsr_report(
  project_directory = NULL,
  pcgr_data = NULL,
  cpsr_config = NULL
)
}
\arguments{
\item{project_directory}{name of project/output directory}

\item{pcgr_data}{Object with underlying knowledge base (PCGR/CPSR)}

\item{cpsr_config}{Object with CPSR configuration parameters
(data directory, virtual panels, sample names, genome assembly etc)}
}
\description{
Function that generates predisposition_report - CPSR
}
