% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main2.R
\name{generate_pcgr_report2}
\alias{generate_pcgr_report2}
\title{Function that generates all contents of the cancer genome report (PCGR)}
\usage{
generate_pcgr_report2(yaml_fname = NULL)
}
\arguments{
\item{yaml_fname}{Name of PCGR configuration file (yaml)}
}
\description{
Function that generates all contents of the cancer genome report (PCGR)
}
