% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predispose.R
\name{generate_predisposition_report}
\alias{generate_predisposition_report}
\title{Function that generates cancer genome report - Tier model pcgr.0}
\usage{
generate_predisposition_report(project_directory, query_vcf2tsv, pcgr_data,
  pcgr_config = NULL, sample_name = "SampleX",
  pcgr_version = "0.1.0", genome_assembly = "grch37")
}
\arguments{
\item{project_directory}{name of project directory}

\item{query_vcf2tsv}{name of gzipped TSV file (vcf2tsv) with annotated query SNVs/InDels}

\item{pcgr_data}{List of data frames with PCGR data annotations}

\item{sample_name}{sample identifier}

\item{pcgr_version}{PCGR software version}

\item{genome_assembly}{human genome assembly version}

\item{pcg_config}{Object with PCGR configuration parameters}
}
\description{
Function that generates cancer genome report - Tier model pcgr.0
}
