% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_report}
\alias{generate_report}
\title{Function that generates all contents of the cancer
genome report (PCGR)}
\usage{
generate_report(yaml_fname = NULL)
}
\arguments{
\item{yaml_fname}{Name of PCGR settings file produced by
pre-reporting Python workflow (yaml)}
}
\description{
Function that generates all contents of the cancer
genome report (PCGR)
}
