% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acmg.R
\name{generate_report_acmg}
\alias{generate_report_acmg}
\title{Function that generates cancer genome report - tier model acmg-like}
\usage{
generate_report_acmg(project_directory, query_vcf2tsv, pcgr_data,
  pcgr_config = NULL, sample_name = "SampleX",
  cna_segments_tsv = NULL, pcgr_version = "0.6.0",
  genome_assembly = "grch37")
}
\arguments{
\item{project_directory}{name of project directory}

\item{query_vcf2tsv}{name of gzipped TSV file (vcf2tsv) with annotated query SNVs/InDels}

\item{pcgr_data}{List of data frames with PCGR data annotations}

\item{sample_name}{sample identifier}

\item{cna_segments_tsv}{name of CNA segments file (tab-separated values)}

\item{pcgr_version}{PCGR software version}

\item{genome_assembly}{human genome assembly version}

\item{pcg_config}{Object with PCGR configuration parameters}
}
\description{
Function that generates cancer genome report - tier model acmg-like
}
