% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{generate_report_data_expression}
\alias{generate_report_data_expression}
\title{Function that generates expression data for PCGR report}
\usage{
generate_report_data_expression(ref_data = NULL, settings = NULL)
}
\arguments{
\item{ref_data}{PCGR reference data object}

\item{settings}{PCGR run/configuration settings}
}
\description{
Function that generates expression data for PCGR report
}
