% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\name{generate_report_data_fusion}
\alias{generate_report_data_fusion}
\title{Function that generates fusion data for PCGR report}
\usage{
generate_report_data_fusion(ref_data = NULL, settings = NULL)
}
\arguments{
\item{ref_data}{PCGR reference data object}

\item{settings}{PCGR run/configuration settings}
}
\description{
Function that generates fusion data for PCGR report
}
