% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msi.R
\name{generate_report_data_msi}
\alias{generate_report_data_msi}
\title{Function that generates MSI prediction data for PCGR report}
\usage{
generate_report_data_msi(sample_calls, pcgr_data, sample_name, pcgr_config)
}
\arguments{
\item{sample_calls}{variant calls subject to mutational signature analysis}

\item{pcgr_data}{object with PCGR annotation data}

\item{sample_name}{sample identifier}

\item{pcgr_config}{Object with PCGR configuration parameters}
}
\description{
Function that generates MSI prediction data for PCGR report
}
