% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational_signatures.R
\name{generate_report_data_rainfall}
\alias{generate_report_data_rainfall}
\title{Function that generates data for rainfall plot (mutation density
along genome, considering SNVs only)}
\usage{
generate_report_data_rainfall(
  variant_set,
  colors = NULL,
  autosomes = F,
  build = "grch37"
)
}
\arguments{
\item{variant_set}{data frame with SNVs/InDels (must contain "CHROM",
"POS","REF","ALT")}

\item{colors}{character vector of six color codes (denoting the
different mutation types)}

\item{autosomes}{logical indicating if plotting should only
consider autosomes}

\item{build}{genome assembly (grch37/grch38)}
}
\description{
Function that generates data for rainfall plot (mutation density
along genome, considering SNVs only)
}
