% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational_signatures.R
\name{generate_report_data_signatures_mp}
\alias{generate_report_data_signatures_mp}
\title{Function that generates mutational signatures data for PCGR report}
\usage{
generate_report_data_signatures_mp(
  callset_snv = NULL,
  ref_data = NULL,
  settings = NULL
)
}
\arguments{
\item{callset_snv}{Somatic callset (SNV)}

\item{ref_data}{PCGR reference data object}

\item{settings}{PCGR configuration settings object}
}
\description{
Function that generates mutational signatures data for PCGR report
}
