% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main2.R
\name{generate_report_data_snv_indel2}
\alias{generate_report_data_snv_indel2}
\title{Function that generates tiered variant sets for SNVs/InDels}
\usage{
generate_report_data_snv_indel2(
  pcg_report = NULL,
  callset = NULL,
  tier_model = "pcgr_acmg"
)
}
\arguments{
\item{pcg_report}{PCGR report object}

\item{callset}{Object with input calls (CNA, SNV/InDel)}

\item{tier_model}{tier model (pcgr_acmg)}
}
\value{
pcg_report_data data frame with all report elements
}
\description{
Function that generates tiered variant sets for SNVs/InDels
}
