% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{generate_tier_tsv_cpsr}
\alias{generate_tier_tsv_cpsr}
\title{Function that generates tiered annotated variant datasets for CPSR}
\usage{
generate_tier_tsv_cpsr(pcg_report, sample_name = "test")
}
\arguments{
\item{pcg_report}{List with tiered variants}
}
\value{
tsv_variants data frame with tier-annotated list of variants for tab-separated output
}
\description{
Function that generates tiered annotated variant datasets for CPSR
}
