% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{get_clinical_associations_cna}
\alias{get_clinical_associations_cna}
\title{Function that retrieves clinical evidence items (CIVIC, CBMDB) for CNA aberrations}
\usage{
get_clinical_associations_cna(onco_ts_sets, pcgr_data, pcgr_config,
  tumor_type_specificity = "any_tumortype")
}
\arguments{
\item{onco_ts_sets}{data frame with lost tumor suppressor genes and gained oncogenes}

\item{pcgr_data}{object with PCGR annotation data}

\item{pcgr_config}{Object with PCGR configuration parameters}

\item{tumor_type_specificity}{run against specific tumor type ('tumor_specific', as specified in pcgr_config) or any tumor type 'any'}
}
\value{
list
}
\description{
Function that retrieves clinical evidence items (CIVIC, CBMDB) for CNA aberrations
}
