% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{get_clinical_associations_snv_indel}
\alias{get_clinical_associations_snv_indel}
\title{Function that retrieves clinical evidence items (CIVIC, CBMDB) for somatic cancer variants}
\usage{
get_clinical_associations_snv_indel(sample_calls, pcgr_data, pcgr_config,
  tumor_type_specificity = "any_tumortype",
  biomarker_mapping_stringency = 1)
}
\arguments{
\item{sample_calls}{data frame with sample variants}

\item{pcgr_data}{object with PCGR annotation data}

\item{pcgr_config}{Object with PCGR configuration parameters}

\item{tumor_type_specificity}{run against specific tumor type ('tumor_specific', as specified in pcgr_config) or any tumor type 'any'}

\item{biomarker_mapping_stringency}{- one of 'exact' (allele-specific) or 'approximate' (codon or exon-level biomarkers)}
}
\value{
list
}
\description{
Function that retrieves clinical evidence items (CIVIC, CBMDB) for somatic cancer variants
}
