% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna.R
\name{get_cna_cytoband}
\alias{get_cna_cytoband}
\title{Function that gets the chromosome bands of copy number segments}
\usage{
get_cna_cytoband(cna_gr, cytoband_gr)
}
\arguments{
\item{cna_gr}{genomic ranges object with copy number aberrations}

\item{cytoband_gr}{genomic ranges object with chromosomal cytobands}

\item{normalization_method}{metod for normalization of context counts (deconstructSigs)}

\item{signatures_limit}{max number of contributing signatures}
}
\description{
Function that gets the chromosome bands of copy number segments
}
