% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna.R
\name{get_cna_overlapping_transcripts}
\alias{get_cna_overlapping_transcripts}
\title{Functions that finds all transcripts that overlap with CNA segments. A
new data frame with one entry per transcript-CNA overlap entry is returned}
\usage{
get_cna_overlapping_transcripts(cna_df, pcgr_data)
}
\arguments{
\item{cna_df}{data frame with copy number segments}

\item{pcgr_data}{PCGR list object with data}
}
\description{
Functions that finds all transcripts that overlap with CNA segments. A
new data frame with one entry per transcript-CNA overlap entry is returned
}
