% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{get_deletion_mechanism}
\alias{get_deletion_mechanism}
\title{Function that annotates underlying mechanism (repeat-mediated, micro-homology etc) of indels}
\usage{
get_deletion_mechanism(vcf_data_df, genome_assembly = "hg19")
}
\arguments{
\item{vcf_data_df}{Data frame with somatic variants}
}
\value{
vcf_data_df
}
\description{
Function that annotates underlying mechanism (repeat-mediated, micro-homology etc) of indels
}
