% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{get_dt_tables}
\alias{get_dt_tables}
\title{Function that gathers data tables on actionable variants
for display in report (tier 1 + tier 2)}
\usage{
get_dt_tables(report = NULL, tier = 1, variant_class = "cna")
}
\arguments{
\item{report}{report object}

\item{tier}{tier level}

\item{variant_class}{cna or snv_indel}
}
\description{
Function that gathers data tables on actionable variants
for display in report (tier 1 + tier 2)
}
