% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsr.R
\name{get_germline_biomarkers}
\alias{get_germline_biomarkers}
\title{Function that retrieves clinical evidence items (CIVIC, CBMDB) for
somatic cancer variants}
\usage{
get_germline_biomarkers(sample_calls, colset = NULL, eitems = NULL)
}
\arguments{
\item{sample_calls}{data frame with germline variant callset from
query sample}

\item{colset}{vector with column names to display for each report element}

\item{eitems}{all clinical evidence items linking germline variants with
impact on therapeutic response/diagnosis/prognosis etc}
}
\value{
list
}
\description{
Function that retrieves clinical evidence items (CIVIC, CBMDB) for
somatic cancer variants
}
