% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{get_insilico_prediction_statistics}
\alias{get_insilico_prediction_statistics}
\title{Function that counts insilico predictions of variant effects (i.e. damaging/tolerated) from dbNSFP}
\usage{
get_insilico_prediction_statistics(sample_calls)
}
\arguments{
\item{sample_calls}{sample calls with dbnsfp annotations}
}
\value{
sample_calls
}
\description{
Function that counts insilico predictions of variant effects (i.e. damaging/tolerated) from dbNSFP
}
