% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsr.R
\name{get_max_rows_pr_datatable}
\alias{get_max_rows_pr_datatable}
\title{Function that gets the maximum number of rows across different
tier data frames in CPSR report}
\usage{
get_max_rows_pr_datatable(cps_report)
}
\arguments{
\item{cps_report}{CPSR report structure with tier data frames}
}
\value{
max_row_nr maximum number of rows
}
\description{
Function that gets the maximum number of rows across different
tier data frames in CPSR report
}
