% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna.R
\name{get_oncogene_tsgene_target_sets}
\alias{get_oncogene_tsgene_target_sets}
\title{Function that detects and assigns oncogenes subject to copy number
amplifications, and tumor suppressor genes subject to homozygous deletions
Also detects other drug targets subject to copy number amplifications}
\usage{
get_oncogene_tsgene_target_sets(
  cna_df,
  transcript_overlap_pct = 100,
  log_r_gain = 0.8,
  log_r_homdel = -0.8,
  tumor_type = "Any",
  pcgr_data = NULL
)
}
\arguments{
\item{cna_df}{data frame with copy number-transcript records}

\item{transcript_overlap_pct}{minimum level of overlap for transcripts (pct)}

\item{log_r_gain}{logR threshold for copy number amplifications}

\item{log_r_homdel}{logR threshold for homozygous deletions}

\item{tumor_type}{type of tumor}

\item{pcgr_data}{PCGR data object}
}
\description{
Function that detects and assigns oncogenes subject to copy number
amplifications, and tumor suppressor genes subject to homozygous deletions
Also detects other drug targets subject to copy number amplifications
}
