% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna.R
\name{get_oncogenic_cna_events}
\alias{get_oncogenic_cna_events}
\title{Get oncogenic copy number events}
\usage{
get_oncogenic_cna_events(cna_df_display = NULL)
}
\arguments{
\item{cna_df_display}{data frame with transcript annotations per
copy number segment}
}
\description{
Function that extracts oncogenic copy number events from
a data frame of annotated transcripts (within copy number segments),
utilizing oncogene/tumor suppressor status and copy number
variant class (gain/loss). This set is used to highlight
copy-altered transcripts that are presumably oncogenic, yet
not actionable \emph{per se}.
}
