% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational_signatures.R
\name{get_prevalent_site_signatures}
\alias{get_prevalent_site_signatures}
\title{Function that retrieves prevalent signatures for a given tumor type/primary site
Data is collected from COSMIC v3.2.}
\usage{
get_prevalent_site_signatures(
  site = "Any",
  custom_collection = NULL,
  pcgr_data = NULL,
  prevalence_pct = 2,
  incl_poss_artifacts = T
)
}
\arguments{
\item{site}{Primary tumor site}

\item{custom_collection}{Custom collection of signatures from COSMIC}

\item{pcgr_data}{PCGR data object}

\item{prevalence_pct}{Minimum prevalence (pct) of signature in cohorts associated with primary site}

\item{incl_poss_artifacts}{logical indicating if artefact signatures are to be included}
}
\description{
Function that retrieves prevalent signatures for a given tumor type/primary site
Data is collected from COSMIC v3.2.
}
