% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{get_proper_maf_alleles}
\alias{get_proper_maf_alleles}
\title{Function that transforms a tier-structured variant data frame
into a MAF-like data frame (for input to 2020plus, MutSigCV)}
\usage{
get_proper_maf_alleles(maf_df, genome_seq, seqinfo)
}
\arguments{
\item{tier_df}{data frame with somatic mutations}
}
\value{
maf_df
}
\description{
Function that transforms a tier-structured variant data frame
into a MAF-like data frame (for input to 2020plus, MutSigCV)
}
