% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{get_tumor_only_filtering_criteria}
\alias{get_tumor_only_filtering_criteria}
\title{Function that generates a string with filtering criteria
for callsets coming from tumor-only sequencing}
\usage{
get_tumor_only_filtering_criteria(conf)
}
\arguments{
\item{conf}{list object with PCGR run configuration settings}
}
\value{
string with filtering criteria
}
\description{
Function that generates a string with filtering criteria
for callsets coming from tumor-only sequencing
}
