% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna.R
\name{get_valid_chromosome_segments}
\alias{get_valid_chromosome_segments}
\title{Function that removes copy number segments that go
beyond chrosomal lengths for the given assembly}
\usage{
get_valid_chromosome_segments(
  cna_segments_df,
  genome_assembly = "grch37",
  bsg = BSgenome.Hsapiens.UCSC.hg19
)
}
\arguments{
\item{cna_segments_df}{genomic segments with CNA}

\item{genome_assembly}{genome assembly (grch37/grch38)}

\item{bsg}{BSgenome object}
}
\description{
Function that removes copy number segments that go
beyond chrosomal lengths for the given assembly
}
