% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_valid_chromosomes}
\alias{get_valid_chromosomes}
\title{Function that check for valid chromosome names in data frame of variants}
\usage{
get_valid_chromosomes(
  vcf_data_df,
  chromosome_column = "CHROM",
  bsg = BSgenome.Hsapiens.UCSC.hg19
)
}
\arguments{
\item{vcf_data_df}{data frame}

\item{chromosome_column}{name of chromosome column}

\item{bsg}{BSGenome object}
}
\value{
vcf_data_df valid data frame with valid mutations
}
\description{
Function that check for valid chromosome names in data frame of variants
}
