% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_stats.R
\name{get_variant_statistics}
\alias{get_variant_statistics}
\title{Function that computes various variant statistics from a data frame
with variant records}
\usage{
get_variant_statistics(var_df = NULL, pct_other_limit = 4)
}
\arguments{
\item{var_df}{data frame with variants}

\item{pct_other_limit}{numeric value specifying the percentage limit
for the 'Other' category}
}
\description{
Function that computes various variant statistics from a data frame
with variant records
}
