% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{grpmax_faf_nc_gnomad}
\alias{grpmax_faf_nc_gnomad}
\title{Function that assigns a maximum value to a variable
(gnomAD_NC_FAF_GRPMAX) reflecting the filter allele frequency
(GrpMAX) for a given variant in the non-cancer
gnomAD subset (v3.1)}
\usage{
grpmax_faf_nc_gnomad(sample_calls)
}
\arguments{
\item{sample_calls}{data frame with sample variant calls}
}
\description{
Function that assigns a maximum value to a variable
(gnomAD_NC_FAF_GRPMAX) reflecting the filter allele frequency
(GrpMAX) for a given variant in the non-cancer
gnomAD subset (v3.1)
}
