% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{het_af_germline_status}
\alias{het_af_germline_status}
\title{Function that assigns a logical (STATUS_LIKELY_GERMLINE_HETEROZYGOUS) reflecting whether
a variant is likely heterozygous (germline) - based on allelic fraction (AF_TUMOR),
presence in gnomAD and dbSNP, and no presence in TCGA and COSMIC}
\usage{
het_af_germline_status(sample_calls)
}
\arguments{
\item{sample_calls}{data frame with sample variant calls}
}
\description{
Function that assigns a logical (STATUS_LIKELY_GERMLINE_HETEROZYGOUS) reflecting whether
a variant is likely heterozygous (germline) - based on allelic fraction (AF_TUMOR),
presence in gnomAD and dbSNP, and no presence in TCGA and COSMIC
}
