% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heterozygous_states}
\alias{heterozygous_states}
\title{VCF encodings of heterozygous variant states}
\format{
A vector of potential VCF encodings of heterozygous variant states
}
\usage{
heterozygous_states
}
\description{
VCF encodings of heterozygous variant states
}
\keyword{datasets}
