% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{hom_af_status}
\alias{hom_af_status}
\title{Function that assigns a logical (STATUS_LIKELY_GERMLINE_HOMOZYGOUS) reflecting whether
a variant is likely homozygous (germline) - based on allelic fraction (VAF_TUMOR)}
\usage{
hom_af_status(sample_calls)
}
\arguments{
\item{sample_calls}{data frame with sample variant calls}
}
\description{
Function that assigns a logical (STATUS_LIKELY_GERMLINE_HOMOZYGOUS) reflecting whether
a variant is likely homozygous (germline) - based on allelic fraction (VAF_TUMOR)
}
