% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{homozygous_states}
\alias{homozygous_states}
\title{VCF encodings of homozygous variant states}
\format{
A vector of potential VCF encodings of homozygous variant states
}
\usage{
homozygous_states
}
\description{
VCF encodings of homozygous variant states
}
\keyword{datasets}
