% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_biomarker_content}
\alias{init_biomarker_content}
\title{Function that initiates report element with biomarker evidence information}
\usage{
init_biomarker_content()
}
\value{
bm_evidence Report structure initialized for biomarker evidence data
}
\description{
Function that initiates report element with biomarker evidence information
}
