% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_cna_vstats}
\alias{init_cna_vstats}
\title{Function that initiates report element with CNA information}
\usage{
init_cna_vstats(rep = NULL)
}
\value{
rep updated PCGR report structure - initialized for CNA content
}
\description{
Function that initiates report element with CNA information
}
