% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_expression_content}
\alias{init_expression_content}
\title{Function that initiates report element with expression information}
\usage{
init_expression_content()
}
\value{
rep Report structure initialized for expression data
}
\description{
Function that initiates report element with expression information
}
