% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_fusion_content}
\alias{init_fusion_content}
\title{Function that initiates report element with fusion information}
\usage{
init_fusion_content()
}
\value{
rep Report structure initialized for fusion data
}
\description{
Function that initiates report element with fusion information
}
