% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_germline_content}
\alias{init_germline_content}
\title{Function that initiates report element with germline variant information (CPSR)}
\usage{
init_germline_content()
}
\value{
rep Report structure initialized for germline data (CPSR)
}
\description{
Function that initiates report element with germline variant information (CPSR)
}
