% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_kataegis_content}
\alias{init_kataegis_content}
\title{Function that initiates report element with kataegis information}
\usage{
init_kataegis_content()
}
\description{
Function that initiates report element with kataegis information
}
