% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_m_signature_content}
\alias{init_m_signature_content}
\title{Function that initiates report element with mutational signatures information}
\usage{
init_m_signature_content()
}
\value{
rep Report structure initialized for signature data
}
\description{
Function that initiates report element with mutational signatures information
}
