% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_pcg_report}
\alias{init_pcg_report}
\title{Function that initates PCGR report object}
\usage{
init_pcg_report(pcgr_config = NULL, sample_name = "SampleX",
  pcgr_version = "0.6.0", genome_assembly = "grch37", class = NULL,
  pcgr_data = NULL, type = "somatic")
}
\arguments{
\item{pcgr_config}{Object with PCGR configuration parameters}

\item{sample_name}{sample identifier}

\item{pcgr_version}{PCGR software version}

\item{genome_assembly}{human genome assembly version}

\item{class}{report analysis section (NULL defaults to full report)}

\item{pcgr_data}{pcgr data object}

\item{cna_segments_tsv}{name of CNA segments file (tab-separated values)}
}
\description{
Function that initates PCGR report object
}
