% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_report}
\alias{init_report}
\title{Function that initiates PCGR/CPSR report object}
\usage{
init_report(yaml_fname = NULL, report_mode = "PCGR")
}
\arguments{
\item{yaml_fname}{Yaml file with configuration settings and paths
to annotated molecular datasets}

\item{report_mode}{Type of report ('PCGR' or 'CPSR')}
}
\description{
Function that initiates PCGR/CPSR report object
}
