% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_report_display_content}
\alias{init_report_display_content}
\title{Function that initiates ranked report display information}
\usage{
init_report_display_content()
}
\value{
rep Report structure initialized for ranked display
}
\description{
Function that initiates ranked report display information
}
