% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_snv_indel_content}
\alias{init_snv_indel_content}
\title{Function that initiates report element with SNV/InDel information}
\usage{
init_snv_indel_content(rep = NULL)
}
\arguments{
\item{rep}{PCGR report structure}
}
\value{
rep updated PCGR report structure - initialized for SNV/InDel content
}
\description{
Function that initiates report element with SNV/InDel information
}
