% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_tmb_content}
\alias{init_tmb_content}
\title{Function that initiates report element with TMB information}
\usage{
init_tmb_content(ref_data = NULL)
}
\arguments{
\item{ref_data}{PCGR reference data list object}
}
\value{
rep TMB report element
}
\description{
Function that initiates report element with TMB information
}
