% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_valuebox_content}
\alias{init_valuebox_content}
\title{Function that initiates report element with value box information}
\usage{
init_valuebox_content()
}
\value{
rep Report structure initialized for value box data
}
\description{
Function that initiates report element with value box information
}
