% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_var_content}
\alias{init_var_content}
\title{Function that initiates report element with variant data}
\usage{
init_var_content()
}
\value{
rep Report structure initialized for variant data
}
\description{
Function that initiates report element with variant data
}
