% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_stats.R
\name{init_vstats_actionable}
\alias{init_vstats_actionable}
\title{Function that initiates report element with actionable
variant statistics information}
\usage{
init_vstats_actionable(vartype = "snv_indel")
}
\arguments{
\item{vartype}{Variant type to initialise statistics for; one of
"snv_indel" (default) or "cna"}
}
\description{
Function that initiates report element with actionable
variant statistics information
}
