% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_stats.R
\name{init_vstats_snv_indel}
\alias{init_vstats_snv_indel}
\title{Function that initiates report element with SNV/InDel statistics information}
\usage{
init_vstats_snv_indel()
}
\description{
Function that initiates report element with SNV/InDel statistics information
}
