% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kataegis.R
\name{kataegis_detect}
\alias{kataegis_detect}
\title{Function that detects kataegis events from a data frame
with genomic cooordinates of mutations}
\usage{
kataegis_detect(
  data,
  sample_id = "sample_id",
  build = "grch37",
  min.mut = 6,
  max.dis = 1000,
  chr = "chr",
  pos = "pos",
  txdb = NULL
)
}
\arguments{
\item{data}{data frame with somatic mutations, as produced by kataegis_input}

\item{sample_id}{sample identifier}

\item{build}{genomoe assembly build}

\item{min.mut}{minimum number of mutations in localized hypermutated region}

\item{max.dis}{maximum distance of kataegis event (basepairs)}

\item{chr}{column name in data that denotes chromosome}

\item{pos}{column name in data that denotes position}

\item{txdb}{transcript database (txdb)}
}
\value{
kataegis_df data frame with potential kataegis events
}
\description{
Function that detects kataegis events from a data frame
with genomic cooordinates of mutations
}
