% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kataegis.R
\name{kataegis_input}
\alias{kataegis_input}
\title{Function that detects kataegis events from a data frame
with genomic cooordinates of mutations}
\usage{
kataegis_input(
  variant_set,
  chr = "chr",
  pos = "pos",
  ref = "ref",
  alt = "alt",
  build = NULL
)
}
\arguments{
\item{variant_set}{data frame with raw set of somatic mutations}

\item{chr}{column name in data that denotes chromosome}

\item{pos}{column name in data that denotes position}

\item{ref}{column name in data that denotes reference allele}

\item{alt}{column name in data that denotes alternate allele}

\item{build}{genome build (grch37 or hg38)}

\item{k}{size of neighbouring sequence context}
}
\description{
Function that detects kataegis events from a data frame
with genomic cooordinates of mutations
}
