% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{list_of_list_to_df}
\alias{list_of_list_to_df}
\title{Helper function to convert list of lists to data.frame}
\usage{
list_of_list_to_df(lst)
}
\arguments{
\item{lst}{List of lists}
}
\description{
Helper function to convert list of lists to data.frame
}
