% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_df}
\alias{list_to_df}
\title{Function that converts a list to a data frame}
\usage{
list_to_df(listfordf)
}
\arguments{
\item{listfordf}{list to be converted}
}
\value{
df
}
\description{
Function that converts a list to a data frame
}
