% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{load_all_eitems}
\alias{load_all_eitems}
\title{Function that loads all evidence items from CIViC and CGI, and
combines them in a unified data.frame}
\usage{
load_all_eitems(eitems_raw = NULL, alteration_type = "MUT", origin = "Somatic")
}
\arguments{
\item{eitems_raw}{raw data frame with evidence items}

\item{alteration_type}{type of alteration ('MUT','MUT_LOF','CNA')}

\item{origin}{variant origin ('Germline','Somatic')}
}
\value{
all_eitems
}
\description{
Function that loads all evidence items from CIViC and CGI, and
combines them in a unified data.frame
}
