% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{load_cpsr_classified_variants}
\alias{load_cpsr_classified_variants}
\title{Function that reads CPSR-classified variants from a TSV file}
\usage{
load_cpsr_classified_variants(
  fname_cpsr_tsv = NA,
  fname_cpsr_yaml = NA,
  cols = NULL,
  ignore_vus = FALSE,
  ref_data = NULL
)
}
\arguments{
\item{fname_cpsr_tsv}{Path to raw input file with CPSR-classified SNVs/InDels}

\item{fname_cpsr_yaml}{Path to YAML configuration file for CPSR analysis}

\item{cols}{column type definitions of raw input file}

\item{ignore_vus}{logical indicating if VUS should be ignored in report}

\item{ref_data}{PCGR reference data object}
}
\description{
Function that reads CPSR-classified variants from a TSV file
}
