% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{load_dna_variants}
\alias{load_dna_variants}
\title{Function that reads and validates CNA or SNV/InDel TSV files
file from PCGR pre-report pipeline}
\usage{
load_dna_variants(fname, cols = NULL, ref_data = NULL)
}
\arguments{
\item{fname}{Path to raw file with DNA aberrations}

\item{cols}{}
}
\description{
Function that reads and validates CNA or SNV/InDel TSV files
file from PCGR pre-report pipeline
}
