% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{load_germline_snv_indel}
\alias{load_germline_snv_indel}
\title{Function that reads and validates an annotated germline SNV/InDel
file from PCGR pre-reporting pipeline}
\usage{
load_germline_snv_indel(fname, ref_data = NULL)
}
\arguments{
\item{fname}{Path to file name}

\item{ref_data}{Object with reference data}
}
\description{
Function that reads and validates an annotated germline SNV/InDel
file from PCGR pre-reporting pipeline
}
