% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_data.R
\name{load_reference_data}
\alias{load_reference_data}
\title{Function that parses and loads reference data from files in the
assembly-specific PCGR bundle directories}
\usage{
load_reference_data(pcgr_db_assembly_dir = NULL, genome_assembly = "grch38")
}
\arguments{
\item{pcgr_db_assembly_dir}{Assembly-specific root path for data bundle}

\item{genome_assembly}{grch37/grch38}
}
\description{
Function that parses and loads reference data from files in the
assembly-specific PCGR bundle directories
}
