% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{load_rna_fusions}
\alias{load_rna_fusions}
\title{Load RNA fusion results}
\usage{
load_rna_fusions(settings = NULL, ref_data = NULL)
}
\arguments{
\item{settings}{PCGR run/configuration settings}

\item{ref_data}{PCGR reference data object}
}
\description{
Load RNA fusion results
}
