% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{load_yaml}
\alias{load_yaml}
\title{Function that loads YAML data with settings and file paths
to annotated molecular profiles}
\usage{
load_yaml(yml_fname, report_mode = "CPSR")
}
\arguments{
\item{yml_fname}{Yaml file with configuration settings and paths
to annotated molecular datasets}

\item{report_mode}{Type of report ('PCGR' or 'CPSR')}
}
\description{
Function that loads YAML data with settings and file paths
to annotated molecular profiles
}
