% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{lof_doc_note}
\alias{lof_doc_note}
\title{Get documentation string for loss-of-function annotation}
\usage{
lof_doc_note()
}
\value{
A documentation string
}
\description{
Get documentation string for loss-of-function annotation
}
