% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{log_var_eitem_stats}
\alias{log_var_eitem_stats}
\title{Function that logs the number of evidence items found, for different
levels of resolution}
\usage{
log_var_eitem_stats(var_eitems = NULL, target_type = "exact")
}
\arguments{
\item{var_eitems}{data frame with variant evidence items}

\item{target_type}{resolution of evidence items}
}
\description{
Function that logs the number of evidence items found, for different
levels of resolution
}
