% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna.R
\name{make_cnaqc_object}
\alias{make_cnaqc_object}
\title{Function that prepares somatic copy number and mutation data for CNAqc plot}
\usage{
make_cnaqc_object(callset_snv = NULL, callset_cna = NULL, settings = NULL)
}
\arguments{
\item{callset_snv}{Somatic SNV callset list object - PCGR}

\item{callset_cna}{Somatic CNA callset list object - PCGR}

\item{settings}{PCGR run configuration settings}
}
\description{
Function that prepares somatic copy number and mutation data for CNAqc plot
}
