% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{map_biomarker_data}
\alias{map_biomarker_data}
\title{Function that maps biomarker identifiers from VCF variant
annotation to full biomarker data tables for display in report}
\usage{
map_biomarker_data(
  varcalls = NULL,
  ref_data = NULL,
  settings = NULL,
  variant_origin = "Somatic",
  vartype = "snv_indel"
)
}
\arguments{
\item{varcalls}{variant calls data frame with biomarker identifiers}

\item{ref_data}{reference data object containing biomarker data}

\item{settings}{PCGR settings object}

\item{variant_origin}{variant origin for filtering biomarker evidence items
(e.g. "Somatic", "Germline", "Any")}

\item{vartype}{variant type for filtering biomarker evidence items
(e.g. "snv_indel", "cna", "fusion")}
}
\value{
data frame with biomarker data for report display
}
\description{
Function that maps biomarker identifiers from VCF variant
annotation to full biomarker data tables for display in report
}
