% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{max_af_onekg}
\alias{max_af_onekg}
\title{Function that assigns a maximum value to a variable (MAX_AF_1KG) reflecting
the maximum allele frequency for a given variant across oneKG populations}
\usage{
max_af_onekg(sample_calls)
}
\arguments{
\item{sample_calls}{data frame with sample variant calls}
}
\description{
Function that assigns a maximum value to a variable (MAX_AF_1KG) reflecting
the maximum allele frequency for a given variant across oneKG populations
}
