% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msi.R
\name{msi_status_is_high}
\alias{msi_status_is_high}
\title{Function that tells if the MSI status of the sample is MSI-high}
\usage{
msi_status_is_high(msi_content = NULL)
}
\arguments{
\item{msi_content}{MSI report element (as returned by
\code{generate_report_data_msi})}
}
\value{
logical
}
\description{
Function that tells if the MSI status of the sample is MSI-high
}
