% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{mutational_signatures_doc_note}
\alias{mutational_signatures_doc_note}
\title{Get documentation string for mutational signatures analysis}
\usage{
mutational_signatures_doc_note()
}
\value{
A documentation string
}
\description{
Get documentation string for mutational signatures analysis
}
