% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oncogenicity_criteria}
\alias{oncogenicity_criteria}
\title{Oncogenicity criteria (ClinGen/CGC/VICC)}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 14 rows and 6 columns.
}
\usage{
oncogenicity_criteria
}
\description{
Oncogenicity criteria (ClinGen/CGC/VICC)
}
\keyword{datasets}
