% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{oncogenicity_doc_note}
\alias{oncogenicity_doc_note}
\title{Get documentation string for oncogenicity annotation}
\usage{
oncogenicity_doc_note()
}
\value{
A documentation string
}
\description{
Get documentation string for oncogenicity annotation
}
