% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oncokb_annotations}
\alias{oncokb_annotations}
\title{Character vector with OncoKB annotations coming from the
MafAnnotator / FusionAnnotator / CnaAnnnotator tools in the PCGR Python workflow.
These annotations are used for variant classification and reporting in PCGR.}
\format{
A character vector with 8 different OncoKB annotations, including:
"ONCOGENICITY_OKB", "MUTATION_EFFECT_OKB" etc
}
\usage{
oncokb_annotations
}
\description{
Character vector with OncoKB annotations coming from the
MafAnnotator / FusionAnnotator / CnaAnnnotator tools in the PCGR Python workflow.
These annotations are used for variant classification and reporting in PCGR.
}
\keyword{datasets}
