% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oncokb_base_api_url}
\alias{oncokb_base_api_url}
\title{OncoKB base API URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
oncokb_base_api_url
}
\description{
OncoKB base API URL
}
\keyword{datasets}
