% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_variants}
\alias{order_variants}
\title{Function that orders genomic aberrations according to order
of chromosomes and chromosomal position}
\usage{
order_variants(vcf_df, chrom_var = "CHROM", pos_var = "POS")
}
\arguments{
\item{vcf_df}{data frame}

\item{chrom_var}{variable name of chromosome in data frame}

\item{pos_var}{variable name for chromosomal position}
}
\value{
vcf_df data frame with ordered mutations
}
\description{
Function that orders genomic aberrations according to order
of chromosomes and chromosomal position
}
