% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_exists}
\alias{pkg_exists}
\title{Does R Package Exist}
\usage{
pkg_exists(p)
}
\arguments{
\item{p}{The R package to check for.}
}
\value{
TRUE if package exists, FALSE otherwise.
}
\description{
Checks if the specified R package exists on the local system.
}
