% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{plot_filtering_stats_exonic}
\alias{plot_filtering_stats_exonic}
\title{Function that generates a pie chart for exonic/non-exonic variant
statistics (for callsets coming from tumor-only sequencing)}
\usage{
plot_filtering_stats_exonic(
  report = NULL,
  plot_margin_top = 50,
  plot_margin_bottom = 20,
  plot_margin_left = 20,
  plot_margin_right = 20,
  font_family = "Helvetica",
  font_size = 15,
  pie_line_width = 3,
  opacity_filtered_categories = 0.4,
  hole_size_pie = 0.4
)
}
\arguments{
\item{report}{list object with PCGR report content}

\item{plot_margin_top}{top margin of the plot}

\item{plot_margin_bottom}{bottom margin of the plot}

\item{plot_margin_left}{left margin of the plot}

\item{plot_margin_right}{right margin of the plot}

\item{font_family}{font family for plot text}

\item{font_size}{font size for plot text}

\item{pie_line_width}{line width for pie chart segments}

\item{opacity_filtered_categories}{opacity for filtered categories}

\item{hole_size_pie}{size of the hole in the pie chart}
}
\value{
filtering_stats list with data frame and plotly pie chart
}
\description{
Function that generates a pie chart for exonic/non-exonic variant
statistics (for callsets coming from tumor-only sequencing)
}
