% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotly_pie_chart}
\alias{plotly_pie_chart}
\title{Plotly Pie Chart - variant statistics}
\usage{
plotly_pie_chart(
  df_variant_stats = NULL,
  category = "CODING_STATUS",
  plot_margin_top = 50,
  plot_margin_bottom = 20,
  plot_margin_left = 20,
  plot_margin_right = 20,
  font_family = "Helvetica",
  font_size = 15,
  pie_line_width = 3,
  opacity_filtered_categories = 0.4,
  hole_size_pie = 0.4
)
}
\arguments{
\item{df_variant_stats}{Data frame with variant statistics}

\item{category}{Category for pie chart (e.g. CODING_STATUS)}

\item{plot_margin_top}{Top margin}

\item{plot_margin_bottom}{Bottom margin}

\item{plot_margin_left}{Left margin}

\item{plot_margin_right}{Right margin}

\item{font_family}{Font family}

\item{font_size}{Font size}

\item{pie_line_width}{Line width for pie chart segments}

\item{opacity_filtered_categories}{Opacity for filtered categories}

\item{hole_size_pie}{Hole size for pie chart (0 to 1)}
}
\value{
Plotly pie chart object
}
\description{
Function that generates a pie chart using plotly
for a given category in a data frame
}
