% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{pon_status}
\alias{pon_status}
\title{Function that assigns a logical (STATUS_PON) reflecting whether
a variant is co-inciding with a variant present in a panel-of-normals database
(PANEL_OF_NORMALS column is TRUE)}
\usage{
pon_status(sample_calls)
}
\arguments{
\item{sample_calls}{data frame with sample variant calls}
}
\description{
Function that assigns a logical (STATUS_PON) reflecting whether
a variant is co-inciding with a variant present in a panel-of-normals database
(PANEL_OF_NORMALS column is TRUE)
}
