% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{popmax_af_gnomad}
\alias{popmax_af_gnomad}
\title{Function that assigns a maximum value to a variable
(gnomAD_AF_POPMAX) reflecting the maximum allele frequency
for a given variant across gnomAD populations}
\usage{
popmax_af_gnomad(sample_calls)
}
\arguments{
\item{sample_calls}{data frame with sample variant calls}
}
\description{
Function that assigns a maximum value to a variable
(gnomAD_AF_POPMAX) reflecting the maximum allele frequency
for a given variant across gnomAD populations
}
