% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{postprocess_vranges_info}
\alias{postprocess_vranges_info}
\title{A function that converts the INFO tags in a VRanges object into a basic types}
\usage{
postprocess_vranges_info(vr)
}
\arguments{
\item{vr}{VRanges object}
}
\value{
vr Vranges object with INFO tags more simply formatted
}
\description{
A function that converts the INFO tags in a VRanges object into a basic types
}
