% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msi.R
\name{predict_msi_status}
\alias{predict_msi_status}
\title{Function that predicts MSI status based on fraction of indels among calls}
\usage{
predict_msi_status(
  variant_set,
  ref_data,
  msi_prediction_model,
  msi_prediction_dataset,
  msi_feature_variance_table = NULL,
  target_size_mb,
  n_calls = NA_integer_,
  sample_name = "Test"
)
}
\arguments{
\item{variant_set}{data frame with somatic mutations/indels}

\item{ref_data}{PCGR reference data object}

\item{msi_prediction_model}{statistical model for MSI prediction}

\item{msi_prediction_dataset}{underlying dataset from TCGA used for
development of statistical classifier}

\item{msi_feature_variance_table}{lookup table (from model training) mapping
mutation-count bins to expected feature SD; used to qualify prediction
confidence. NULL if not available.}

\item{target_size_mb}{size of targeted genomic region (coding)}

\item{n_calls}{number of variants passed to the classifier}

\item{sample_name}{name of sample}
}
\value{
msi_data
}
\description{
Function that predicts MSI status based on fraction of indels among calls
}
