% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_logical_vars}
\alias{recode_logical_vars}
\title{Function that assigns proper logical values to logical entries
read from VCF file (TRUE = True (VCF), FALSE = False (VCF))}
\usage{
recode_logical_vars(vcf_data_df)
}
\arguments{
\item{vcf_data_df}{variant data frame}
}
\description{
Function that assigns proper logical values to logical entries
read from VCF file (TRUE = True (VCF), FALSE = False (VCF))
}
