% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_table.R
\name{render_oncogenicity_cell}
\alias{render_oncogenicity_cell}
\title{Style oncogenicity cell background based on value
(Generated by Claude Opus 4.6 with some manual tweaks)}
\usage{
render_oncogenicity_cell(data)
}
\arguments{
\item{data}{The main data frame (passed via closure)}
}
\value{
An HTML span element with background color corresponding
to the oncogenicity category, using colors defined in
color_palette$oncogenicity.
}
\description{
Style oncogenicity cell background based on value
(Generated by Claude Opus 4.6 with some manual tweaks)
}
