% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsr.R
\name{retrieve_secondary_calls}
\alias{retrieve_secondary_calls}
\title{Function that retrieves variants in genes recommended for secondary
(secondary) findings}
\usage{
retrieve_secondary_calls(calls, umls_map)
}
\arguments{
\item{calls}{data frame with variants in genes recommended for
secondary findings reporting}

\item{umls_map}{data frame with UMLS phenotype terms}
}
\description{
Function that retrieves variants in genes recommended for secondary
(secondary) findings
}
