% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\name{rna_fusion_recurrence_mitdb}
\alias{rna_fusion_recurrence_mitdb}
\title{Function that checks for recurrence of RNA fusions in Mitelman database
and annotates with clinical and cytogenetic features}
\usage{
rna_fusion_recurrence_mitdb(query_fusions = NULL, ref_data = NULL)
}
\arguments{
\item{query_fusions}{data frame with RNA fusion calls}

\item{ref_data}{PCGR reference data bundle (list)}
}
\value{
data frame with RNA fusion calls annotated with Mitelman database features
}
\description{
Function that checks for recurrence of RNA fusions in Mitelman database
and annotates with clinical and cytogenetic features
}
