% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{set_report_metadata}
\alias{set_report_metadata}
\title{Function that set PCGR report metadata}
\usage{
set_report_metadata(
  config,
  pcgr_data,
  virtual_panel_id = "-1",
  report_type = NULL,
  custom_bed = NULL
)
}
\arguments{
\item{config}{PCGR config object}

\item{pcgr_data}{PCGR data list object}

\item{virtual_panel_id}{identifier for virtual panel}

\item{report_type}{type of report ('germline' or 'somatic')}

\item{custom_bed}{custom BED file with target loci for screening}
}
\value{
report_metadata Object with PCGR report metadata
}
\description{
Function that set PCGR report metadata
}
