% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_chromosomal_segments}
\alias{sort_chromosomal_segments}
\title{Function that sorts chromosomal segments according to chromosome
and chromosomal start/end position}
\usage{
sort_chromosomal_segments(
  df,
  chromosome_column = "CHROM",
  start_segment = "POS",
  end_segment = "POS"
)
}
\arguments{
\item{df}{data frame with chromosome and start + end segment}

\item{chromosome_column}{name of column for chromosome name is sigven}

\item{start_segment}{name of column that indicates start of
chromosomal segment}

\item{end_segment}{name of column that indicates end of chromosomal segment}
}
\value{
df_final data frame with sorted chromosomal segments
}
\description{
Function that sorts chromosomal segments according to chromosome
and chromosomal start/end position
}
