% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_stats.R
\name{stats_report_cna}
\alias{stats_report_cna}
\title{Function that computes various variant statistics for CNAs
from a callset object}
\usage{
stats_report_cna(callset = NULL)
}
\arguments{
\item{callset}{list object with callset data (CNA)}
}
\description{
Function that computes various variant statistics for CNAs
from a callset object
}
