% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_stats.R
\name{stats_report_fusion}
\alias{stats_report_fusion}
\title{Function that computes various variant statistics for fusions
from a callset object}
\usage{
stats_report_fusion(callset = NULL)
}
\arguments{
\item{callset}{list object with callset data (fusions)}
}
\description{
Function that computes various variant statistics for fusions
from a callset object
}
