% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_stats.R
\name{stats_report_germline}
\alias{stats_report_germline}
\title{Function that generate stats for a germline variant callset,
including number of variants and number of variants with
evidence items for each BM evidence type}
\usage{
stats_report_germline(var_df = NULL, vartype = "snv_indel")
}
\arguments{
\item{var_df}{data.frame with germline variant data (SNVs/InDels)}

\item{vartype}{type of variant ('snv_indel')}
}
\description{
Function that generate stats for a germline variant callset,
including number of variants and number of variants with
evidence items for each BM evidence type
}
