% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_stats.R
\name{stats_report_snv_indel}
\alias{stats_report_snv_indel}
\title{Function that computes various variant statistics for
SNVs/InDels from a callset object}
\usage{
stats_report_snv_indel(callset = NULL)
}
\arguments{
\item{callset}{list object with callset data (SNVs/InDels)}
}
\description{
Function that computes various variant statistics for
SNVs/InDels from a callset object
}
