% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{structure_var_eitems}
\alias{structure_var_eitems}
\title{Function that structures variant evidence items according
to strength of evidence}
\usage{
structure_var_eitems(var_eitems, annotation_tags, alteration_type = "MUT")
}
\arguments{
\item{var_eitems}{variant evidence items}

\item{annotation_tags}{annotation tags to include for display}

\item{alteration_type}{type of alteration ('MUT','CNA')}
}
\description{
Function that structures variant evidence items according
to strength of evidence
}
