% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsr.R
\name{summary_donut_chart}
\alias{summary_donut_chart}
\title{Function that makes a piechart showing the number of variants at
each significance level}
\usage{
summary_donut_chart(variants_tsv, plot_type = "ClinVar")
}
\arguments{
\item{variants_tsv}{data frame with variants in predisposition_genes}

\item{plot_type}{ClinVar or Other}
}
\description{
Function that makes a piechart showing the number of variants at
each significance level
}
