% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predispose.R
\name{summary_findings_cpsr}
\alias{summary_findings_cpsr}
\title{Function that generates summary findings for CPSR}
\usage{
summary_findings_cpsr(pcg_report)
}
\arguments{
\item{pcg_report}{report object}
}
\description{
Function that generates summary findings for CPSR
}
