% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_display_cols}
\alias{table_display_cols}
\title{HTML Table display columns}
\format{
An object of class \code{list} of length 3.
}
\usage{
table_display_cols
}
\description{
HTML Table display columns
}
\keyword{datasets}
