% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{targeted_drugs_pr_ttype}
\alias{targeted_drugs_pr_ttype}
\title{Function that retrieves targeted drugs for a given tumor type/primary site}
\usage{
targeted_drugs_pr_ttype(ttype, pcgr_data, ignore_on_label_early_phase = T)
}
\arguments{
\item{ttype}{primary site/tumor type}

\item{pcgr_data}{PCGR data object}

\item{ignore_on_label_early_phase}{ignore early phase drugs (on label)}
}
\description{
Function that retrieves targeted drugs for a given tumor type/primary site
}
