% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{targeted_drugs_summarise}
\alias{targeted_drugs_summarise}
\title{Function that summarises available targeted drugs for a given target}
\usage{
targeted_drugs_summarise(
  candidate_drugs = NULL,
  link_label = "DRUGS_ON_LABEL",
  indication_label = "DRUGS_ON_LABEL_INDICATIONS"
)
}
\arguments{
\item{candidate_drugs}{data frame with candidate drugs}

\item{link_label}{column name for list of aggregated drugs pr. target}

\item{indication_label}{column name for drug indications}
}
\description{
Function that summarises available targeted drugs for a given target
}
