% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tier_af_distribution}
\alias{tier_af_distribution}
\title{Function that plots a histogram of the the variant allelic
support (tumor) - grouped by tiers}
\usage{
tier_af_distribution(tier_df, bin_size = 0.1)
}
\arguments{
\item{tier_df}{data frame with somatic mutations}

\item{bin_size}{size of bins for allelic frequency}
}
\value{
p geom_histogram plot from ggplot2
}
\description{
Function that plots a histogram of the the variant allelic
support (tumor) - grouped by tiers
}
