% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tier_to_maf}
\alias{tier_to_maf}
\title{Function that transforms a tier-structured variant data frame into a MAF-like data frame (for input to 2020plus, MutSigCV)}
\usage{
tier_to_maf(tier_df)
}
\arguments{
\item{tier_df}{data frame with somatic mutations}
}
\value{
maf_df
}
\description{
Function that transforms a tier-structured variant data frame into a MAF-like data frame (for input to 2020plus, MutSigCV)
}
