% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{tmb_doc_note}
\alias{tmb_doc_note}
\title{Get documentation string for tumor mutational burden (TMB) annotation}
\usage{
tmb_doc_note()
}
\value{
A documentation string
}
\description{
Get documentation string for tumor mutational burden (TMB) annotation
}
