% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tumor_sites}
\alias{tumor_sites}
\title{Character vector of tumor site names used in PCGR/CPSR reports}
\format{
A character vector of 31 tumor site names (e.g. "Any", "Adrenal Gland",
"Biliary Tract"). Used to constrain tumor site selection in configuration.
}
\usage{
tumor_sites
}
\description{
Character vector of tumor site names used in PCGR/CPSR reports
}
\keyword{datasets}
