% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{twohit_doc_note}
\alias{twohit_doc_note}
\title{Get documentation string for two-hit detection &
VAF consistency assessment}
\usage{
twohit_doc_note()
}
\value{
A documentation string
}
\description{
Get documentation string for two-hit detection &
VAF consistency assessment
}
