% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{update_pcg_report}
\alias{update_pcg_report}
\title{Function that initates PCGR report object}
\usage{
update_pcg_report(pcg_report, report_data,
  analysis_element = "snv_indel")
}
\arguments{
\item{pcg_report}{PCGR final report}

\item{report_data}{Object with PCGR report data}

\item{analysis_element}{section of PCGR report}
}
\description{
Function that initates PCGR report object
}
