% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{update_report}
\alias{update_report}
\title{Function that updates a PCGR/CPSR report object structure}
\usage{
update_report(report, report_data, a_elem = "snv_indel")
}
\arguments{
\item{report}{PCGR/CPSR report}

\item{report_data}{Object with report contents}

\item{a_elem}{section of PCGR report}
}
\description{
Function that updates a PCGR/CPSR report object structure
}
