% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{upset_plot_tumor_only}
\alias{upset_plot_tumor_only}
\title{Function that makes an upset calls for germline-filtered variants
classification procedure}
\usage{
upset_plot_tumor_only(upset_data)
}
\arguments{
\item{upset_data}{unfiltered calls (germline + somatic)}
}
\value{
p
}
\description{
Function that makes an upset calls for germline-filtered variants
classification procedure
}
