% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{vaf_plot}
\alias{vaf_plot}
\title{Function that generates a VAF distribution plot for a given PCGR report object}
\usage{
vaf_plot(report = NULL, font_size = 12, font_family = "Helvetica")
}
\arguments{
\item{report}{PCGR report object}

\item{font_size}{font size for plot text elements}

\item{font_family}{font family for plot text elements}
}
\value{
vaf_plot_plotly plotly object with VAF distribution plot
}
\description{
Function that generates a VAF distribution plot for a given PCGR report object
}
