% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variant_db_url}
\alias{variant_db_url}
\title{List of URLS and variant identifiers for variant/gene/protein domain databases}
\format{
A data.frame with 6 rows and 5 columns that indicates URL's for various variant/gene databases
and how to use PCGR annotation columns to generate variant links
\itemize{
\item \emph{name} - Name encoding for variant/gene database
\item \emph{group_by_var} - Which column should be used for grouping
\item \emph{url_prefix} - URL prefix
\item \emph{link_key_var} - Which column to be used as the key value in link
\item \emph{link_display_var} - Which column to be used as the display variable in link
}
}
\usage{
variant_db_url
}
\description{
List of URLS and variant identifiers for variant/gene/protein domain databases
}
\keyword{datasets}
