% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{virtual_panel_display_html}
\alias{virtual_panel_display_html}
\title{Function that makes a HTML display of virtual gene panel}
\usage{
virtual_panel_display_html(gene_df)
}
\arguments{
\item{gene_df}{data frame with genes targeted in virtual panel}
}
\description{
Function that makes a HTML display of virtual gene panel
}
