% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{write_report_flexdb_html}
\alias{write_report_flexdb_html}
\title{Function that writes contents of PCGR object to flexdashboard
HTML reports}
\usage{
write_report_flexdb_html(report = NULL)
}
\arguments{
\item{report}{List object with all report data, settings etc.}
}
\description{
Function that writes contents of PCGR object to flexdashboard
HTML reports
}
