% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{write_report_tsv}
\alias{write_report_tsv}
\title{Function that writes contents of PCGR object to flexdashboard
HTML reports}
\usage{
write_report_tsv(report = NULL, variant_type = "snv_indel")
}
\arguments{
\item{report}{List object with all report data, settings etc.}

\item{variant_type}{character indicating variant type for output TSV,
i.e. 'snv_indel' or 'cna_gene'}
}
\description{
param report List object with all report data, settings etc.
}
\details{
export
Function that writes contents of PCGR object to a TSV file
}
