% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_read_support}
\alias{add_read_support}
\title{Function that adds read support (depth, allelic fraction) for tumor and normal}
\usage{
add_read_support(vcf_data_df, tumor_dp_tag = "_na", tumor_af_tag = "_na",
  normal_dp_tag = "_na", normal_af_tag = "_na", call_conf_tag = "_na")
}
\arguments{
\item{vcf_data_df}{}
}
\value{
vcf_data_df
}
\description{
Function that adds read support (depth, allelic fraction) for tumor and normal
}
