% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_cancer_gene_evidence}
\alias{append_cancer_gene_evidence}
\title{Function that appends cancer gene evidence links}
\usage{
append_cancer_gene_evidence(vcf_data_df = NULL, pcgr_data = NULL)
}
\arguments{
\item{vcf_data_df}{Data frame of sample variants from VCF}

\item{pcgr_data}{PCGR reference data bundle object}
}
\value{
vcf_data_df
}
\description{
Function that appends cancer gene evidence links
}
