% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{append_cancer_gene_evidence}
\alias{append_cancer_gene_evidence}
\title{Function that appends cancer gene evidence links}
\usage{
append_cancer_gene_evidence(var_data_df = NULL, ref_data = NULL)
}
\arguments{
\item{var_data_df}{Data frame of sample variants from VCF}

\item{ref_data}{PCGR/CPSR reference data object}
}
\value{
var_data_df
}
\description{
Function that appends cancer gene evidence links
}
