% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{append_drug_var_link}
\alias{append_drug_var_link}
\title{Function that adds link to targeted drugs (on and off-label)
for a list of variants and associated targeted}
\usage{
append_drug_var_link(var_df, primary_site = "Breast", ref_data = NULL)
}
\arguments{
\item{var_df}{data frame with variants}

\item{primary_site}{primary tumor sample site}

\item{ref_data}{PCGR/CPSR reference data list object}
}
\description{
Function that adds link to targeted drugs (on and off-label)
for a list of variants and associated targeted
}
