% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_gwas_citation_phenotype}
\alias{append_gwas_citation_phenotype}
\title{Function that adds GWAS citation/phenotype to GWAS hit found through PCGR annotation}
\usage{
append_gwas_citation_phenotype(vcf_data_df = NULL, ref_data = NULL)
}
\arguments{
\item{vcf_data_df}{Data frame of sample variants from VCF}

\item{ref_data}{PCGR/CPSR reference data object}
}
\value{
vcf_data_df
}
\description{
Function that adds GWAS citation/phenotype to GWAS hit found through PCGR annotation
}
