% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_tcga_var_link}
\alias{append_tcga_var_link}
\title{Function that adds TCGA annotations (cohort, frequency etc.) to variant identifiers}
\usage{
append_tcga_var_link(var_df, linktype = "dbsource")
}
\arguments{
\item{var_df}{data frame with variants}

\item{linktype}{type of link}
}
\value{
var_df
}
\description{
Function that adds TCGA annotations (cohort, frequency etc.) to variant identifiers
}
