% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{assign_germline_popfreq_status}
\alias{assign_germline_popfreq_status}
\title{Function that sets gnomAD_AF_ABOVE_TOLERATED to TRUE for variants
if any gnomAD population frequency exceeds max_tolerated_af}
\usage{
assign_germline_popfreq_status(sample_calls, max_af = 0.01)
}
\arguments{
\item{sample_calls}{data frame with variants}

\item{max_af}{max tolerated germline allele frequency}
}
\value{
sample_calls
}
\description{
Function that sets gnomAD_AF_ABOVE_TOLERATED to TRUE for variants
if any gnomAD population frequency exceeds max_tolerated_af
}
