% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{assign_somatic_classification}
\alias{assign_somatic_classification}
\title{Function that assigns a SOMATIC_CLASSIFICATION to variants
based on evidence found in variant set,
potentially limited by user-defined options}
\usage{
assign_somatic_classification(sample_calls, config)
}
\arguments{
\item{sample_calls}{data frame with variants}

\item{config}{configuration object}
}
\value{
sample_calls
}
\description{
Function that assigns a SOMATIC_CLASSIFICATION to variants
based on evidence found in variant set,
potentially limited by user-defined options
}
