% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_read_support}
\alias{filter_read_support}
\title{Function that filters variant set on depth and allelic fraction
according to settings provided by user (tumor and control)}
\usage{
filter_read_support(vcf_df, config = NULL)
}
\arguments{
\item{vcf_df}{data frame with variants}

\item{config}{list with PCGR configuration settings}
}
\value{
vcf_df
}
\description{
Function that filters variant set on depth and allelic fraction
according to settings provided by user (tumor and control)
}
