% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_report}
\alias{generate_report}
\title{Function that generates cancer genome report}
\usage{
generate_report(project_directory, query_vcf, pcgr_data, pcgr_config = NULL,
  sample_name = "SampleX", cna_segments_tsv = NULL,
  pcgr_version = "0.6.0", genome_assembly = "grch37")
}
\arguments{
\item{project_directory}{name of project directory}

\item{query_vcf}{name of VCF file with annotated query SNVs/InDels}

\item{pcgr_data}{List of data frames with PCGR data annotations}

\item{sample_name}{sample identifier}

\item{cna_segments_tsv}{name of CNA segments file (tab-separated values)}

\item{pcgr_version}{PCGR software version}

\item{pcg_config}{Object with PCGR configuration parameters}

\item{configuration_file}{Configuration file (TOML) for germline variant exclusion criteria (tumor_only mode)}
}
\value{
p
}
\description{
Function that generates cancer genome report
}
