% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_report_data_cna}
\alias{generate_report_data_cna}
\title{Function that annotates CNV segment files}
\usage{
generate_report_data_cna(
  cna_segments_tsv,
  pcgr_data,
  sample_name,
  pcgr_config,
  transcript_overlap_pct = 100
)
}
\arguments{
\item{cna_segments_tsv}{CNV file name with chromosomal log(2)-ratio segments}

\item{pcgr_data}{object with PCGR annotation data}

\item{sample_name}{sample identifier}

\item{pcgr_config}{Object with PCGR configuration parameters}

\item{transcript_overlap_pct}{required aberration overlap fraction
(percent) for reported transcripts (default 100 percent)}
}
\description{
Function that annotates CNV segment files
}
