% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msi.R
\name{generate_report_data_msi}
\alias{generate_report_data_msi}
\title{Function that generates MSI prediction data for PCGR report}
\usage{
generate_report_data_msi(variant_set, ref_data = NULL, settings = NULL)
}
\arguments{
\item{variant_set}{variant calls subject to MSI classification}

\item{ref_data}{PCGR reference data object}

\item{settings}{PCGR run configuration settings}
}
\description{
Function that generates MSI prediction data for PCGR report
}
