% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msi.R
\name{generate_report_data_msi}
\alias{generate_report_data_msi}
\title{Function that generates MSI prediction data for PCGR report}
\usage{
generate_report_data_msi(sample_calls, pcgr_data, pcgr_version,
  sample_name, pcgr_config, genome_seq, genome_assembly)
}
\arguments{
\item{sample_calls}{variant calls subject to mutational signature analysis}

\item{pcgr_data}{object with PCGR annotation data}

\item{pcgr_version}{PCGR software version}

\item{sample_name}{sample identifier}

\item{pcgr_config}{Object with PCGR configuration parameters}

\item{genome_seq}{BSgenome object}

\item{genome_assembly}{human genome assembly version}
}
\description{
Function that generates MSI prediction data for PCGR report
}
