% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational_signatures.R
\name{generate_report_data_signatures}
\alias{generate_report_data_signatures}
\title{Function that generates mutational signatures data for PCGR report}
\usage{
generate_report_data_signatures(
  variant_set = NULL,
  ref_data = NULL,
  settings = NULL,
  n_bootstrap_iterations = 200,
  sig_contribution_cutoff = 0.01
)
}
\arguments{
\item{variant_set}{Somatic callset (SNV)}

\item{ref_data}{PCGR reference data object}

\item{settings}{PCGR run/configuration settings}

\item{n_bootstrap_iterations}{Number of bootstrap iterations for signature analysis}

\item{sig_contribution_cutoff}{Minimal signature contribution for reporting}
}
\description{
Function that generates mutational signatures data for PCGR report
}
