% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational_signatures.R
\name{generate_report_data_signatures_mp}
\alias{generate_report_data_signatures_mp}
\title{Function that generates mutational signatures data for PCGR report}
\usage{
generate_report_data_signatures_mp(
  vcf_fname,
  pcgr_data,
  sample_name,
  pcgr_config,
  type_specific = T
)
}
\arguments{
\item{vcf_fname}{VCF file processed with PCGR annotation pipeline -
possibly filtered for depth/allelic fraction}

\item{pcgr_data}{object with PCGR annotation data}

\item{sample_name}{sample identifier}

\item{pcgr_config}{Object with PCGR configuration parameters}

\item{type_specific}{logical indicating if all reference signatures are to be
included (F) rather than those known to be prevalent in the tumor (T)}
}
\description{
Function that generates mutational signatures data for PCGR report
}
