% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_report_data_snv_indel}
\alias{generate_report_data_snv_indel}
\title{Function that generates tiered report data for SNVs/InDels}
\usage{
generate_report_data_snv_indel(sample_calls, pcgr_data, sample_name = NULL,
  pcgr_config = NULL, callset = "somatic calls",
  biomarker_mapping_stringency = 1, bsg = BSgenome.Hsapiens.UCSC.hg19,
  genome_assembly = "hg19")
}
\arguments{
\item{sample_calls}{data frame with list of variant calls}

\item{pcgr_data}{List of data frames with PCGR annotations}

\item{pcgr_config}{object with pcgr configurations (msi, mutationalsignatures etc)}

\item{callset}{type of calls}

\item{biomarker_mapping_stringency}{quality level for biomarkers}

\item{bsg}{genome sequence object (BSgenome.Hsapiens.UCSC)}

\item{genome_assembly}{grch37/grch38}

\item{sample_id}{sample identifier}
}
\value{
report_data data frame with all report elements
}
\description{
Function that generates tiered report data for SNVs/InDels
}
