% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational_burden.R
\name{generate_report_data_tmb}
\alias{generate_report_data_tmb}
\title{Function that tiered variant sets for SNVs/InDels}
\usage{
generate_report_data_tmb(sample_calls, pcgr_data, sample_name, pcgr_config)
}
\arguments{
\item{sample_calls}{variant calls subject to mutational signature analysis}

\item{pcgr_data}{object with PCGR annotation data}

\item{sample_name}{sample identifier}

\item{pcgr_config}{Object with PCGR configuration parameters}
}
\value{
pcg_report_data data frame with all report elements
}
\description{
Function that tiered variant sets for SNVs/InDels
}
