% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational_burden.R
\name{generate_report_data_tmb}
\alias{generate_report_data_tmb}
\title{Function that reads TSV file with TMB estimates from sample}
\usage{
generate_report_data_tmb(settings = NULL)
}
\arguments{
\item{settings}{PCGR run/configuration settings}
}
\value{
tmb_estimate
}
\description{
Function that reads TSV file with TMB estimates from sample
}
