% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{generate_report_data_tumor_only}
\alias{generate_report_data_tumor_only}
\title{Function that generates germline-filtered callset and PCGR report statistics for tumor-only callsets}
\usage{
generate_report_data_tumor_only(unfiltered_sample_calls, pcgr_data,
  pcgr_version, sample_name, pcgr_config, genome_seq, genome_assembly)
}
\arguments{
\item{unfiltered_sample_calls}{variant calls subject to mutational signature analysis}

\item{pcgr_data}{object with PCGR annotation data}

\item{pcgr_version}{PCGR software version}

\item{sample_name}{sample identifier}

\item{pcgr_config}{Object with PCGR configuration parameters}

\item{genome_seq}{BSgenome object}

\item{genome_assembly}{human genome assembly version}
}
\description{
Function that generates germline-filtered callset and PCGR report statistics for tumor-only callsets
}
