% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/main2.R
\name{generate_report_data_tumor_only}
\alias{generate_report_data_tumor_only}
\title{Function that generates germline-filtered callset and PCGR
report statistics for a given tumor-only callsets}
\usage{
generate_report_data_tumor_only(
  unfiltered_sample_calls,
  sample_name,
  pcgr_config
)

generate_report_data_tumor_only(
  unfiltered_sample_calls,
  sample_name,
  pcgr_config
)
}
\arguments{
\item{unfiltered_sample_calls}{variant calls}

\item{sample_name}{sample identifier}

\item{pcgr_config}{Object with PCGR configuration parameters}
}
\description{
Function that generates germline-filtered callset and PCGR
report statistics for a given tumor-only callsets

Function that generates germline-filtered callset and PCGR
report statistics for a given tumor-only callsets
}
