% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_boxes.R
\name{generate_report_data_value_box}
\alias{generate_report_data_value_box}
\title{Function that generates value box data for PCGR report}
\usage{
generate_report_data_value_box(pcg_report, pcgr_data, sample_name, pcgr_config)
}
\arguments{
\item{pcg_report}{object with existing PCGR report data elements}

\item{pcgr_data}{object with PCGR annotation data}

\item{sample_name}{sample identifier}

\item{pcgr_config}{Object with PCGR configuration parameters}
}
\description{
Function that generates value box data for PCGR report
}
