% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_tier_tsv}
\alias{generate_tier_tsv}
\title{Function that generates dense and tiered annotated variant datasets}
\usage{
generate_tier_tsv(variant_set, config, annotation_tags, sample_name = "test")
}
\arguments{
\item{variant_set}{List with tiered variants}

\item{config}{PCGR configuration settings}

\item{annotation_tags}{List with display columns}

\item{sample_name}{Sample identifier}
}
\value{
tsv_variants data frame with tier-annotated list of
variants for tab-separated output
}
\description{
Function that generates dense and tiered annotated variant datasets
}
