% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{germline_filter_levels}
\alias{germline_filter_levels}
\title{Data frame with germline filtering criteria}
\format{
A character vector listing all germline filtering criteria
applied on input callsets (SNVs/InDels) in tumor-only mode
}
\usage{
germline_filter_levels
}
\description{
Data frame with germline filtering criteria
}
\keyword{datasets}
