% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_calls}
\alias{get_calls}
\title{Function that reads a VCF from pcgr pipeline}
\usage{
get_calls(vcf_gz_file, pcgr_data, pcgr_config, sample_id = NULL,
  genome_assembly = "hg19")
}
\arguments{
\item{vcf_gz_file}{Bgzipped VCF file}

\item{pcgr_data}{Lisf ot data frames with PCGR data annotations}

\item{sample_id}{Sample identifier}
}
\value{
vcf_data_df
}
\description{
Function that reads a VCF from pcgr pipeline
}
