% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna.R
\name{get_cna_cytoband}
\alias{get_cna_cytoband}
\title{Function that gets the chromosome bands of copy number segments}
\usage{
get_cna_cytoband(cna_df, pcgr_data = NULL)
}
\arguments{
\item{cna_df}{genomic ranges object with copy number aberrations}

\item{pcgr_data}{pcgr data bundle object}
}
\description{
Function that gets the chromosome bands of copy number segments
}
