% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsr.R
\name{get_insilico_prediction_statistics}
\alias{get_insilico_prediction_statistics}
\title{Function that counts insilico predictions of variant effects
(i.e. damaging/tolerated) from dbNSFP}
\usage{
get_insilico_prediction_statistics(cpg_calls)
}
\arguments{
\item{cpg_calls}{sample calls with dbNSFP annotations}
}
\value{
cpg_calls
}
\description{
Function that counts insilico predictions of variant effects
(i.e. damaging/tolerated) from dbNSFP
}
