% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_valid_chromosomes}
\alias{get_valid_chromosomes}
\title{Checks for valid chromosome names in data frame of variants}
\usage{
get_valid_chromosomes(vcf_data_df, chromosome_column = "CHROM", bsg = NULL)
}
\arguments{
\item{vcf_data_df}{data frame}

\item{chromosome_column}{name of chromosome column}

\item{bsg}{BSGenome object}
}
\value{
vcf_data_df valid data frame with valid mutations
}
\description{
Checks for valid chromosome names in data frame of variants
}
