% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_valid_chromosomes}
\alias{get_valid_chromosomes}
\title{Function that check for valid mutations/chromosomes in input}
\usage{
get_valid_chromosomes(vcf_data_df, chromosome_column = "CHROM",
  bsg = BSgenome.Hsapiens.UCSC.hg19)
}
\arguments{
\item{vcf_data_df}{data frame}

\item{chromosome_column}{name of columns for which string replace is to be performed}

\item{pattern}{pattern to replace}
}
\value{
vcf_data_df_valid data frame with valid mutations
}
\description{
Function that check for valid mutations/chromosomes in input
}
