% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_report}
\alias{init_report}
\title{Function that initiates PCGR/CPSR report object}
\usage{
init_report(
  config = NULL,
  sample_name = "SampleX",
  class = NULL,
  pcgr_data = NULL,
  pcgr_version = "dev",
  cpsr_version = "dev",
  type = "somatic",
  virtual_panel_id = "-1",
  custom_bed = NULL
)
}
\arguments{
\item{config}{Object with configuration parameters}

\item{sample_name}{sample identifier}

\item{class}{report analysis section (NULL defaults to full report)}

\item{pcgr_data}{PCGR data bundle}

\item{pcgr_version}{PCGR software version}

\item{cpsr_version}{CPSR software version}

\item{type}{Type of report ('somatic' or 'germline')}

\item{virtual_panel_id}{identifier(s) for virtual panel id}

\item{custom_bed}{custom BED file with target loci for screening}
}
\description{
Function that initiates PCGR/CPSR report object
}
