% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_tmb_content}
\alias{init_tmb_content}
\title{Function that initiates report element with TMB information}
\usage{
init_tmb_content(tcga_tmb = NULL, config = NULL)
}
\arguments{
\item{tcga_tmb}{data frame with TMB distribution in TCGA samples}

\item{config}{PCGR configuration object}
}
\value{
rep TMB report element
}
\description{
Function that initiates report element with TMB information
}
