% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{load_dna_variants}
\alias{load_dna_variants}
\title{Function that reads and validates CNA or SNV/InDel TSV files
file from PCGR/CPSR pre-report (Python) pipeline}
\usage{
load_dna_variants(
  fname = NA,
  cols = NULL,
  ref_data = NULL,
  vartype = "snv_indel",
  primary_site = "Any",
  retained_info_tags = "None",
  variant_origin = "Somatic"
)
}
\arguments{
\item{fname}{Path to raw input file with DNA aberrations (PCGR/CPSR)}

\item{cols}{column type definitions of raw input file}

\item{ref_data}{reference data object}

\item{vartype}{type of DNA aberrations ('snv_indel','cna')}

\item{primary_site}{primary site of tumor}

\item{retained_info_tags}{VCF INFO tags to be retained in output (SNVs/InDels)}

\item{variant_origin}{Germline/Somatic}
}
\description{
Function that reads and validates CNA or SNV/InDel TSV files
file from PCGR/CPSR pre-report (Python) pipeline
}
