% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{load_expression_outliers}
\alias{load_expression_outliers}
\title{Load expression outlier results}
\usage{
load_expression_outliers(
  settings = NULL,
  ref_data = NULL,
  percentile_cutoff_high = 95,
  percentile_cutoff_low = 5,
  z_score_cutoff = 1.5
)
}
\arguments{
\item{settings}{PCGR run/configuration settings}

\item{ref_data}{PCGR reference data object}

\item{percentile_cutoff_high}{numeric, percentile cutoff for high expression}

\item{percentile_cutoff_low}{numeric, percentile cutoff for low expression}

\item{z_score_cutoff}{numeric, z-score cutoff for expression outliers}
}
\description{
Load expression outlier results
}
