% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{load_somatic_cna}
\alias{load_somatic_cna}
\title{Function that reads and validates a fully annotated CNA file from PCGR
pre-report pipeline}
\usage{
load_somatic_cna(fname, ref_data = NULL, settings = NULL)
}
\arguments{
\item{fname}{Path to file name}

\item{ref_data}{Object with reference data}

\item{settings}{Object with PCGR report configuration}
}
\description{
Function that reads and validates a fully annotated CNA file from PCGR
pre-report pipeline
}
