% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{load_somatic_snv_indel}
\alias{load_somatic_snv_indel}
\title{Function that reads and validates a fully annotated somatic SNV/InDel
file from PCGR pre-report pipeline}
\usage{
load_somatic_snv_indel(fname)
}
\arguments{
\item{fname}{Path to file name}
}
\description{
Function that reads and validates a fully annotated somatic SNV/InDel
file from PCGR pre-report pipeline
}
