% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{load_somatic_snv_indel}
\alias{load_somatic_snv_indel}
\title{Function that reads and validates an annotated somatic SNV/InDel
file from PCGR pre-reporting pipeline}
\usage{
load_somatic_snv_indel(fname = NA, ref_data = NULL, settings = NULL)
}
\arguments{
\item{fname}{Path to file name}

\item{ref_data}{Object with reference data}

\item{settings}{Object with PCGR report configuration}
}
\description{
Function that reads and validates an annotated somatic SNV/InDel
file from PCGR pre-reporting pipeline
}
