% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna.R
\name{plot_cna_segments}
\alias{plot_cna_segments}
\title{Plot allele-specific copy number segments}
\usage{
plot_cna_segments(chrom_coordinates = NULL, cna_segments = NULL)
}
\arguments{
\item{chrom_coordinates}{data frame with assembly-specific chromosome coordinate data (length etc)}

\item{cna_segments}{data frame with annotated copy number segments}
}
\description{
Function that plots allele-specific copy number segments
(minor + total allele copies)
}
