% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational_burden.R
\name{plot_tmb_primary_site_tcga}
\alias{plot_tmb_primary_site_tcga}
\title{Function that makes a plot with TMB boxplots for TCGA cohorts, highlighting
the TMB estimate for a given sample and the cohort/primary site of interest}
\usage{
plot_tmb_primary_site_tcga(
  tcga_tmb,
  p_site = "Liver",
  tmb_estimate = 5,
  algorithm = "all_coding"
)
}
\arguments{
\item{tcga_tmb}{data frame with TMB estimates for TCGA samples}

\item{p_site}{primary tumor_site (sample)}

\item{tmb_estimate}{estimate of mutational burden (sample)}

\item{algorithm}{type of TMB algorithm}
}
\description{
Function that makes a plot with TMB boxplots for TCGA cohorts, highlighting
the TMB estimate for a given sample and the cohort/primary site of interest
}
