% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msi.R
\name{predict_msi_status}
\alias{predict_msi_status}
\title{Function that predicts MSI status based on fraction of indels among calls}
\usage{
predict_msi_status(vcf_data_df, simpleRepeats_gr, windowMasker_gr,
  msi_prediction_model, indelFracPlot_template,
  bsg = BSgenome.Hsapiens.UCSC.hg19, genome_assembly = "hg19",
  sample_name = "Test")
}
\arguments{
\item{vcf_data_df}{data frame with somatic mutations/indels}

\item{simpleRepeats_gr}{Genomic Ranges object with sequence repeats}

\item{windowMasker_gr}{Genomic Ranges object with}

\item{msi_prediction_model}{statistical model for MSI prediction/classification}

\item{bsg}{BSgenome data object in Biostrings objects (e.g. BSgenome.Hsapiens.UCSC.hg19)}

\item{genome_assembly}{hg19/hg38}

\item{sample_name}{name of sample}

\item{indel_frac_plot_tcga}{distribution plot of indel fraction for TCGA samples with known MSI status}
}
\value{
msi_data
}
\description{
Function that predicts MSI status based on fraction of indels among calls
}
