% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msi.R
\name{predict_msi_status}
\alias{predict_msi_status}
\title{Function that predicts MSI status based on fraction of indels among calls}
\usage{
predict_msi_status(
  vcf_data_df,
  pcgr_data,
  msi_prediction_model,
  msi_prediction_dataset,
  target_size_mb,
  sample_name = "Test"
)
}
\arguments{
\item{vcf_data_df}{data frame with somatic mutations/indels}

\item{pcgr_data}{object with PCGR datasets}

\item{msi_prediction_model}{statistical model for MSI prediction}

\item{msi_prediction_dataset}{underlying dataset from TCGA used for
development of statistical classifier}

\item{target_size_mb}{size of targeted genomic region (coding)}

\item{sample_name}{name of sample}
}
\value{
msi_data
}
\description{
Function that predicts MSI status based on fraction of indels among calls
}
