% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{qc_var_eitems}
\alias{qc_var_eitems}
\title{Function that makes a quality control check of evidence items assigned
to variants}
\usage{
qc_var_eitems(var_eitems = NULL, marker_type = "codon")
}
\arguments{
\item{var_eitems}{variant-evidence items}

\item{marker_type}{type of biomarker}
}
\description{
Function that makes a quality control check of evidence items assigned
to variants
}
