% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational_signatures.R
\name{signature_contributions_single_sample}
\alias{signature_contributions_single_sample}
\title{Function that retrieves relative estimates of known somatic signatures from a single tumor}
\usage{
signature_contributions_single_sample(mut_data, sample_name,
  normalization_method = "default", cosmic_signatures_aetiologies = NULL,
  signatures_limit = 6, bsg = BSgenome.Hsapiens.UCSC.hg19)
}
\arguments{
\item{mut_data}{data frame with somatic mutations (VCF_SAMPLE_ID, CHROM, POS, REF, ALT)}

\item{sample_name}{sample name}

\item{normalization_method}{metod for normalization of context counts (deconstructSigs)}

\item{signatures_limit}{max number of contributing signatures}

\item{cosmic_cancertypes_aetiologies}{list of known signaturea and associated etiologies/cancertypes}
}
\description{
Function that retrieves relative estimates of known somatic signatures from a single tumor
}
