% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{update_report}
\alias{update_report}
\title{Function that initates PCGR report object}
\usage{
update_report(report, report_data, a_elem = "snv_indel")
}
\arguments{
\item{report}{PCGR final report}

\item{report_data}{Object with PCGR report data}

\item{a_elem}{section of PCGR report}
}
\description{
Function that initates PCGR report object
}
