% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{variant_stats_report}
\alias{variant_stats_report}
\title{Function that generate snv/indel + coding/noncoding stats for
a given variant set}
\usage{
variant_stats_report(calls, name = "v_stat")
}
\arguments{
\item{calls}{data frame with variants in predisposition_genes}

\item{name}{type of variant group}
}
\description{
Function that generate snv/indel + coding/noncoding stats for
a given variant set
}
