% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{variant_stats_report}
\alias{variant_stats_report}
\title{Function that generate snv/indel + coding/noncoding + tier stats for
a given variant set}
\usage{
variant_stats_report(calls = NULL, name = "vstats", vartype = "snv_indel")
}
\arguments{
\item{calls}{data frame with variants}

\item{name}{type of variant statistic}

\item{vartype}{type of variant ('snv_indel', 'cna')}
}
\description{
Function that generate snv/indel + coding/noncoding + tier stats for
a given variant set
}
