% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{write_report_output}
\alias{write_report_output}
\title{Function that writes contents of PCGR object to various output formats
(Rmarkdown/flexdashboard HTML reports, JSON, tab-separated etc)}
\usage{
write_report_output(
  project_directory,
  report,
  sample_name,
  genome_assembly,
  tier_model = "pcgr_acmg",
  output_format = "html",
  flexdb = F
)
}
\arguments{
\item{project_directory}{working directory}

\item{report}{List object with all report data (PCGR/CPSR)}

\item{sample_name}{sample name}

\item{genome_assembly}{genome assembly (grch37/grch38)}

\item{tier_model}{type of tier model}

\item{output_format}{contents/file format of output
(html/json/tsv/cna_tsv etc)}

\item{flexdb}{logical indicating if HTML output should be dashboard}
}
\description{
Function that writes contents of PCGR object to various output formats
(Rmarkdown/flexdashboard HTML reports, JSON, tab-separated etc)
}
