% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{write_report_tsv}
\alias{write_report_tsv}
\title{Function that writes contents of PCGR object to various output formats
(Rmarkdown/flexdashboard HTML reports, JSON, tab-separated etc)}
\usage{
write_report_tsv(report = NULL, output_type = "snv_indel")
}
\arguments{
\item{report}{List object with all report data, settings etc.}

\item{output_type}{character indicating output type for TSV,
i.e. 'snv_indel', 'snv_indel_unfiltered', 'cna_gene', or 'msigs'}
}
\description{
param report List object with all report data (PCGR/CPSR), settings etc.
param tier_model type of tier model
param output_format contents/file format of output
(html/json/tsv/cna_tsv etc)
param flexdb logical indicating if HTML output should be dashboard
}
\details{
export
Function that writes contents of PCGR object to flexdashboard
HTML reports

param report List object with all report data, settings etc.

export
Function that writes contents of PCGR object to a TSV file
}
